% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_height}
\alias{age_height}
\title{Age vs. height for private and rural school children}
\format{
A data frame with 18 rows and 2 variables:
\describe{
\item{Months}{Child's age in months}
\item{Height}{Child's height in cm}
...
}
}
\source{
Understanding robust and exploratory data analysis, by D.C. Hoaglin,
F. Mosteller and J.W. Tukey. (page 135)
}
\usage{
age_height
}
\description{
The data were reproduced from Hoaglin et al.'s book which  were originally
sourced from Bernard G. Greenberg (1953) in the American Journal of Public
Health (vol 43, pp. 692-699). The dataset tabulate children's
height and weight from urban private and rural public schools.
}
\keyword{datasets}
