% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_plot_vardecomp.R
\name{.eda_plot_vardecomp}
\alias{.eda_plot_vardecomp}
\title{Variability decomposition plots}
\usage{
.eda_plot_vardecomp(
  dat,
  response,
  type = "boxpnt",
  input = "nway",
  eff = NULL,
  rotate = FALSE,
  padding = 0.1,
  show.resp = FALSE,
  outliers = TRUE,
  label = FALSE,
  order = TRUE,
  cex.txt = 1,
  lim = NULL,
  overlap = c("stack", "overplot", "jitter"),
  pch = 16,
  p.col = "grey50",
  p.fill = "grey80",
  size = 1,
  alpha = 0.5,
  grey = 0.6,
  title = "Variability decomposition",
  ...
)
}
\arguments{
\item{dat}{A data frame in long form, containing the data to be plotted.}

\item{response}{A character string specifying the name of the response
variable column in the \code{dat} data frame.}

\item{type}{A character string specifying the type of plot to generate.
Must be either \code{"boxpnt"} or \code{"box"}.}

\item{input}{A character string. \code{"reg"} = bivariate model input.
\code{"nway"} = univariate model or N-way table input.}

\item{eff}{A list of effect values. Required when \code{input = "nway"}.}

\item{rotate}{Logical. If \code{TRUE}, rotates the plot orientation.}

\item{padding}{Numeric. Controls padding for plot limits.}

\item{show.resp}{Logical. If \code{TRUE}, includes a boxplot for the response
variable.}

\item{outliers}{Logical. If \code{TRUE}, outliers are displayed in boxplots.}

\item{label}{Logical. Controls whether labels are displayed.}

\item{order}{Logical. Controls ordering (likely of factors or effects).}

\item{cex.txt}{Numeric. Controls text size.}

\item{lim}{Numeric. Explicit limits for the plot axes.}

\item{overlap}{Character. Specifies how to handle overlapping points,
must be one of \code{"stack"}, \code{"overplot"}, or \code{"jitter"}.}

\item{pch}{Point symbol type. Only applicable if \code{type = "boxpnt"}.}

\item{p.col}{Point border color. Only applicable if \code{type = "boxpnt"}.}

\item{p.fill}{Point fill color. Only applicable if \code{type = "boxpnt"}.}

\item{size}{Point size. Only applicable if \code{type = "boxpnt"}.}

\item{alpha}{Transparency level for points (0 = transparent, 1 = opaque).}

\item{grey}{Numeric. Controls grayscale coloring for plot elements
and axes.}

\item{title}{Plot title. If title is to be omitted, set to \code{NULL}.}

\item{...}{Additional arguments passed to underlying plotting functions.}
}
\value{
Primarily called for producing a plot.
}
\description{
This is an internal helper function used to generate plots
that visualize the effects and residuals from the analysis of n-way tables or
a regression model.
The function is not intended for direct use by the end-user.
}
\keyword{internal}
