% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_3pt.R
\name{eda_3pt}
\alias{eda_3pt}
\title{3-point summary plot}
\usage{
eda_3pt(
  dat,
  x,
  y,
  px = 1,
  py = 1,
  tukey = TRUE,
  base = exp(1),
  axes = TRUE,
  pch = 21,
  equal = TRUE,
  p.col = "grey50",
  p.fill = "grey80",
  size = 0.8,
  alpha = 0.7,
  xlab = NULL,
  ylab = NULL,
  dir = TRUE,
  grey = 0.6,
  ...
)
}
\arguments{
\item{dat}{Dataframe.}

\item{x}{Variable assigned the x axis.}

\item{y}{Variable assigned the y axis.}

\item{px}{Power transformation to apply to the x-variable.}

\item{py}{Power transformation to apply to the y-variable.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation).}

\item{base}{Base used with the log() function if \code{px} or
\code{py} is \code{0}.}

\item{axes}{Boolean determining if axes are to be drawn.}

\item{pch}{Point symbol type.}

\item{equal}{Boolean determining if axes lengths should match (i.e. square
plot).}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point size (0-1).}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only applicable
if \code{rgb()} is not used to define point colors.}

\item{xlab}{X label for output plot.}

\item{ylab}{Y label for output plot.}

\item{dir}{Boolean indicating if suggested ladder of power direction should
be displayed.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{...}{Other parameters passed to the graphics::plot function.}
}
\value{
Generates a plot and returns a list with the following named
components:

\itemize{
\item \code{hsrtio}: The ratio between both slopes. A value close to one
suggests that no transformation is needed.
\item \code{xmed}: The x-coordinate values for the three summary points.
\item \code{ymed}: The y-coordinate values for the three summary points.}
}
\description{
\code{eda_3pt} splits the data into 3 groups (whose summary
locations are defined by their respective medians), and two half slopes
linking the groups. The function will return a scatter plot showing the
half-slopes in red solid lines. The solid grey slope linking both tail-end
groups shows the desired shape for both half-slopes. The goal is to have the
two halve slopes line up as closely as possible to the solid grey slope via
re-expression techniques when seeking a linear relationship between both
variables.

The function will also return the half-slopes ratio \code{hsrtio} and the
direction of re-expression for both X and Y values on the ladder of powers.
}
\details{
Computes the three-point summary originally defined in Tukey's EDA
book (see reference).
}
\examples{

hsratio <- eda_3pt(cars, speed, dist)
hsratio <- eda_3pt(cars, speed, dist, py = 1/3, ylab=expression("Dist"^{1/3}))

#' # This next example models gas consumption as a function of engine displacement.
# It applies a transformation to both variables via the px and py arguments.
eda_3pt(mtcars, disp, mpg,  px = -1/3, py = -1,
       ylab = "gal/mi", xlab = expression("Displacement"^{-1/3}))

}
\references{
\itemize{
\item Velleman, P. F., and D. C. Hoaglin. 1981. Applications, Basics and Computing of Exploratory Data Analysis. Boston: Duxbury Press.
\item D. C. Hoaglin, F. Mosteller, and J. W. Tukey. 1983. Understanding Robust and Exploratory Data Analysis. Wiley.
\item Tukey, John W. 1977. Exploratory Data Analysis. Addison-Wesley.}
}
