% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_add.R
\name{eda_add}
\alias{eda_add}
\title{Add graphical EDA elements to existing plot}
\usage{
eda_add(
  x,
  pch = 24,
  p.col = "darkred",
  p.fill = "yellow",
  lty = 1,
  l.col = "darkred"
)
}
\arguments{
\item{x}{Object of class \code{eda_rline}.}

\item{pch}{Point symbol type.}

\item{p.col}{Point color passed to \code{col}.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{lty}{Line type.}

\item{l.col}{Line color.}
}
\value{
Returns the \code{eda_rline} intercept and slope.

\itemize{
  \item \code{a}: Intercept
  \item \code{b}: Slope}
}
\description{
\code{eda_add}  adds graphical EDA elements to a scatter plot.
  Currently only adds \code{eda_rline} fit and points.
}
\details{
This function adds an eda_rline slope and 3-pt summary points to an
  existing scatter plot.

  \cr See the \href{../articles/polish.html}{Resistant Line article} for a
  detailed breakdown of the resistant line technique.
}
\examples{

eda_lm(mtcars, x = wt, y = mpg)
Mr <- eda_rline(mtcars, x=wt, y=mpg)
eda_add(Mr, l.col = "blue")
}
