% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_dens.R
\name{eda_dens}
\alias{eda_dens}
\title{Overlapping density distributions for two variables}
\usage{
eda_dens(
  x,
  y,
  fac = NULL,
  p = 1L,
  tukey = FALSE,
  base = exp(1),
  fx = NULL,
  fy = NULL,
  grey = 0.6,
  col = "red",
  show.par = TRUE,
  alpha = 0.4,
  xlab = NULL,
  ylab = NULL,
  xaxis = NULL,
  switch = FALSE,
  legend = TRUE,
  ...
)
}
\arguments{
\item{x}{Vector for first variable, or a dataframe.}

\item{y}{Vector for second variable, or column defining the continuous
variable if \code{x} is a dataframe.}

\item{fac}{Column defining the categorical variable if \code{x} is a dataframe.}

\item{p}{Power transformation to apply to both sets of values.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation).}

\item{base}{Base used with the log() function if \code{p = 0}.}

\item{fx}{Formula to apply to x variable. This is computed after any
transformation is applied to the x variable.}

\item{fy}{Formula to apply to y variable. This is computed after any
transformation is applied to the y variable.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{col}{Fill color for second density distribution.}

\item{show.par}{Boolean determining if parameters such as power
transformation or formula should be displayed.}

\item{alpha}{Fill transparency (0 = transparent, 1 = opaque). Only applicable
if \code{rgb()} is not used to define fill colors.}

\item{xlab}{X variable label. Ignored if \code{x} is a dataframe. This is
not the x-axis label!}

\item{ylab}{Y variable label. Ignored if \code{x} is a dataframe. This is
not the y-axis label!}

\item{xaxis}{Label for x-axis. Defaults to \code{"Value"}.}

\item{switch}{Boolean determining if the axes should be swapped. Only applies
to dataframe input. Ignored if vectors are passed to the function.}

\item{legend}{Boolean determining if a legend should be added to the plot.}

\item{...}{Arguments passed to the \code{stats::density()} function.}
}
\value{
Does not return a value.
}
\description{
\code{eda_dens} generates overlapping density distributions for
two variables.
}
\details{
This function will generate overlapping density plots with the first
variable assigned a grey color and the second variable assigned the default
red color.
}
\examples{

 # Passing data as two separate vector objects
 set.seed(207)
 x <- rbeta(1000,2,8)
 y <- x * 1.5 + 0.1
 eda_dens(x, y)

 # Passing data as a dataframe
 dat <- data.frame(val = c(x, y),
                   grp = c(rep("x", length(x)), rep("y", length(y))))
 eda_dens(dat, val, grp)
}
