% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_fval.R
\name{eda_fval}
\alias{eda_fval}
\title{Calculate Quantile f-values}
\usage{
eda_fval(x, q.type = 5)
}
\arguments{
\item{x}{Vector whose f-values are to be computed.}

\item{q.type}{Quantile algorithm.}
}
\value{
A vector of f-values in the same order as vector \code{x}.
}
\description{
\code{eda_fval} Generates f-values for quantile functions
}
\details{
For each value i in \code{x}, \code{eda_fval} computes the
 cumulative probability associated with i. It provides a measure of the
 position of a data point relative to the rest of the data scaled to the
 range [0,1]. This is essentially the fraction, f-value,  of the dataset
 that lies below i. This fraction is sometimes reported as the
 probability or cumulative frequency. \cr
 \cr
 Different algorithms are used to compute the f-value. \code{eda_fval}
 offers as optional algorithm type \code{4} through \code{6} documented
 in \code{stats::quantile}. The algorithms used are: \cr \cr
  \itemize{
    \item \code{4}: f = i / n,
    \item \code{5}: f = (i - 0.5) / n,
    \item \code{6}: f = i / (n + 1),
    \item \code{7}: f = (i  - 1) / (n - 1),
    \item \code{8}: f = (i - 1/3) / (n + 1/3),
    \item \code{9}: f = (i - 3/8) / (n + 1/4).}

  Where f is the fraction of values that lies below index i, and
  n is the total number values.
}
\examples{

 set.seed(321)
 z <- runif(10, 1, 20)

 # William Cleveland's f-values algorithm
 eda_fval(z)

 # Algorithm used by the stats::quantile() function
 eda_fval(z, q.type = 7)
}
\references{
\itemize{
 \item John M. Chambers, William S. Cleveland, Beat Kleiner, Paul A. Tukey.
  Graphical Methods for Data Analysis (1983)}
}
