% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_lm.R
\name{eda_lm}
\alias{eda_lm}
\title{Regression plot (with optional LOESS fit)}
\usage{
eda_lm(
  dat,
  x,
  y,
  xlab = NULL,
  ylab = NULL,
  px = 1,
  py = 1,
  tukey = FALSE,
  base = exp(1),
  ...
)
}
\arguments{
\item{dat}{Dataframe.}

\item{x}{Column assigned to the x axis.}

\item{y}{Column assigned to the y axis.}

\item{xlab}{X label for output plot.}

\item{ylab}{Y label for output plot.}

\item{px}{Power transformation to apply to the x-variable.}

\item{py}{Power transformation to apply to the y-variable.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation).}

\item{base}{Base used with the log() function if \code{px} or
\code{py} is \code{0}.}

\item{...}{
  Arguments passed on to \code{\link{.eda_plot_xy}}
  \describe{
    \item{\code{raw_tick}}{Logical. If \code{TRUE}, original (untransformed) equally
spaced tick values are displayed on the re-expressed axes.}
    \item{\code{xlim}}{X-axis range.}
    \item{\code{ylim}}{Y-axis range.}
    \item{\code{show.par}}{Logical; whether to display plot parameter summary on the
plot. Currently only applies to regression model input.}
    \item{\code{reg}}{Logical; whether to fit and display a regression line.}
    \item{\code{poly}}{Integer; regression model polynomial degree (defaults to 1 for
linear model).}
    \item{\code{robust}}{Logical; if \code{TRUE}, uses robust regression (\code{MASS::rlm}).}
    \item{\code{rlm.d}}{List; parameters for \code{MASS::rlm}, (e.g., \code{list(psi =
"psi.bisquare")}).}
    \item{\code{w}}{Optional numeric vector of weights for regression.}
    \item{\code{lm.col}}{Regression line color.}
    \item{\code{lm.lw}}{Numeric; Regression line width.}
    \item{\code{lm.lty}}{Numeric; Regression line type.}
    \item{\code{sd}}{Logical; whether to show ±1 SD lines.}
    \item{\code{mean.l}}{Logical; whether to show x and y mean reference lines.}
    \item{\code{asp}}{Logical; whether to preserve the aspect ratio (ignored if
\code{square = FALSE}).}
    \item{\code{square}}{Logical; whether to create a square plotting window.}
    \item{\code{grey}}{Numeric between \code{0-1}; controls grayscale background elements
(\code{0 = black}, \code{1 = white}).}
    \item{\code{pch}}{Integer; point symbol.}
    \item{\code{p.col}}{Point border color.}
    \item{\code{p.fill}}{Point fill color.}
    \item{\code{size}}{Point size.}
    \item{\code{alpha}}{Point transparency level (0 = 100\\\% transparent, 1 = 100\\\% opaque).}
    \item{\code{q}}{Logical; whether to draw inner quantile boxes (quantile shading).}
    \item{\code{q.type}}{Integer; type of quantile calculation (see \code{quantile}).}
    \item{\code{inner}}{Numeric; defines the inner fraction of values to highlight
with quantile shading.}
    \item{\code{qcol}}{Fill color of quantile shading.}
    \item{\code{loe}}{Logical; whether to plot loess smooth line.}
    \item{\code{loe.lw}}{Numeric; Loess smooth line width.}
    \item{\code{loe.col}}{Loess smooth color.}
    \item{\code{loe.lty}}{Numeric; Loess smooth line type.}
    \item{\code{loess.d}}{List; parameters for \code{loess.smooth}, e.g., \code{list(span = 0.7,
degree = 1)}.}
    \item{\code{stats}}{Logical; if \code{TRUE}, displays model statistics (R², β, p-value).}
    \item{\code{stat.size}}{Text size for \code{stats} plot display.}
    \item{\code{hline}}{Numeric; location(s) of additional horizontal reference lines.
Can be passed via  the \code{c()} function.}
    \item{\code{vline}}{Numeric; location(s) of additional vertical reference lines.
Can be passed via the \code{c()} function.}
    \item{\code{plot}}{Logical. Generates a plot if \code{TRUE}.}
  }}
}
\value{
Returns a list of class \code{eda_lm}. Output includes the following
if \code{reg = TRUE}. Returns \code{NULL} otherwise.

\itemize{
\item \code{data}: Input data table with residuals
\item \code{residuals}: Regression model residuals
\item \code{a}: Intercept
\item \code{b}: Polynomial coefficient(s)
\item \code{fitted.values}: Fitted values
\item \code{x}: x variable
\item \code{x_lab}: x label}
}
\description{
\code{eda_lm} generates a scatter and EDA enhanced regression
plot.
}
\details{
The function will plot a regression line and, if requested, a loess
fit. The function adopts the least squares fitting technique by default. It
defaults to a first order polynomial fit. The polynomial order can be
specified via the \code{poly} argument.
\cr\cr
The plot displays the +/- 1 standard deviations as dashed lines. In
theory, if both x and y values follow a perfectly Normal distribution,
roughly 68 percent of the points should fall in between these lines.
\cr \cr
The true 68 percent of values can be displayed as a shaded region by
setting \code{q=TRUE}. It uses the \code{quantile} function to compute
the upper and lower bounds defining the inner 68 percent of values. If the
data follow a Normal distribution, the grey rectangle edges should coincide
with the +/- 1SD dashed lines.
If you wish to show the interquartile ranges (IQR) instead of the inner
68 percent of values, simply set \code{inner = 0.5}.
\cr \cr
The function offers the option to re-express the values via the \code{px} and
\code{py} arguments. But note that if the re-expression produces \code{NaN}
values (such as if a negative value is logged) those points will be
removed from the plot. This will result in fewer observations being
plotted. If observations are removed as a result of a re-expression, a
warning message will be displayed in the console.
The re-expression powers are shown in the upper right side of the plot. To
suppress the display of the re-expressions set \code{show.par = FALSE}.
\cr\cr
If the \code{robust} argument is set to TRUE, \code{MASS}'s
built-in robust fitting model, \code{rlm}, is used to fit the regression
line to the data. \code{rlm} arguments can be passed as a list via the
\code{rlm.d} argument.
}
\examples{

# Add a regular (OLS) regression model and loess smooth to the data
eda_lm(mtcars, wt, mpg, loe = TRUE)

# Add the inner 68\% quantile to compare the true 68\% of data to the SD
eda_lm(mtcars, wt, mpg, loe = TRUE, q = TRUE)

# Show the IQR box
eda_lm(mtcars, wt, mpg, loe = TRUE, q = TRUE, sd = FALSE, inner = 0.5)

# Fit an OLS to income for Female vs Male
inc <- read.csv("https://mgimond.github.io/ES218/Data/Income_education.csv")
eda_lm(inc, x=B20004013, y = B20004007, xlab = "Female", ylab = "Male",
            loe = TRUE)

# Add the inner 68\% quantile to compare the true 68\% of data to the SD
eda_lm(inc, x = B20004013, y = B20004007, xlab = "Female", ylab = "Male",
            q = TRUE)

# Apply a transformation to x and y axes: x -> 1/3 and y -> log
eda_lm(inc, x = B20004013, y = B20004007, px = 1/3, py = 0, loe = TRUE,
            xlab = expression(("Female income") ^ frac(1,3)),
            ylab = "log(Male income)")

# You can opt to show the original values on scaled axes
eda_lm(inc, x = B20004013, y = B20004007, px = 1/3, py = 0, loe = TRUE,
            xlab = "Female", ylab = "Male", raw_tick = TRUE)

# Fit a second order polynomial
eda_lm(mtcars, hp, mpg, poly = 2)

# Fit a robust regression model
eda_lm(mtcars, hp, mpg, robust = TRUE, poly = 2)
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link[stats]{loess.smooth}},
\code{\link{.eda_plot_xy}}
}
