% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_lsum.R
\name{eda_lsum}
\alias{eda_lsum}
\title{Tukey's letter value summaries}
\usage{
eda_lsum(x, l = 5, all = TRUE)
}
\arguments{
\item{x}{Vector}

\item{l}{Number of levels}

\item{all}{generate upper, lower and mid summaries if TRUE or
just generate mid summaries if FALSE}
}
\description{
\code{eda_lsum} The letter value summary was introduced by John Tukey
 and extends the boxplot's 5 number summary by exploring the symmetry
 of the batch for depth levels other than the half (median) or the
 fourth (quartiles)
}
\details{
Visit this  [link](https://mgimond.github.io/ES218/Week08b.html) for
  more information on the letter values summaries.
}
\examples{
x <- c(22, 8, 11, 3, 26, 1, 14, 18, 20, 25, 24)
eda_lsum(x)
}
\references{
Exploratory Data Analysis, John Tukey, 1973.
}
