% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_matlong.R
\name{eda_matlong}
\alias{eda_matlong}
\title{Convert 2-way data tables Between Long and Matrix Formats}
\usage{
eda_matlong(
  data,
  row,
  col,
  response,
  direction = c("long_to_matrix", "matrix_to_long")
)
}
\arguments{
\item{data}{A data frame in long format (for \code{direction = "long_to_matrix"})
or a matrix (for \code{direction = "matrix_to_long"}).}

\item{row}{A \strong{bare (unquoted)} variable name from \code{data} (or a desired
column name for the output) representing the row identifiers in the matrix.}

\item{col}{A \strong{bare (unquoted)} variable name from \code{data} (or a desired
column name for the output) representing the column identifiers in the
matrix.}

\item{response}{A \strong{bare (unquoted)} variable name from \code{data} (or a
desired column name for the output) representing the values in the matrix.}

\item{direction}{A character string specifying the direction of the
conversion. Must be one of \code{"long_to_matrix"} or \code{"matrix_to_long"}.
Defaults to \code{"long_to_matrix"}.}
}
\value{
If \code{direction = "long_to_matrix"}, returns a matrix. If \code{direction =
  "matrix_to_long"}, returns a data frame in long format.
}
\description{
Reshapes a 2-way data table from long format to a matrix,
or a matrix back to a data frame in long format.
}
\examples{
eda_matlong(edtts2.12, col = year, row = institution, response = perc)
}
