% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_mean_sweep.R
\name{eda_mean_sweep}
\alias{eda_mean_sweep}
\title{Exploratory ANOVA Decomposition}
\usage{
eda_mean_sweep(
  data,
  ...,
  max_order = 1,
  nesting = NULL,
  p = 1,
  tukey = FALSE,
  base = exp(1)
)
}
\arguments{
\item{data}{A data frame containing the response and factor variables.}

\item{...}{Unquoted variable names: The first variable must be the numeric
response, followed by one or more factor variables.}

\item{max_order}{An integer specifying the maximum order of
interaction effects to include. Main effects are always computed if factors
are provided.
\describe{
\item{\code{max_order = 1}}{ (default) Only main effects are calculated and included
in the decomposition.}
\item{\code{max_order = 2}}{Main effects and all two-way interactions among the
specified factors are included.}
\item{\code{max_order = k}}{Main effects and all interaction terms up to order \code{k}
are included.}
}}

\item{nesting}{A list of character vectors specifying nested relationships.
Each element should be a pair like \code{c("Parent", "Child")} indicating
that \code{Child} is nested within \code{Parent}. If provided, the function
will automatically reorder the sweeping sequence to ensure that parent
factors are swept before their nested children.}

\item{p}{Numeric. A power transformation to apply to the response variable
before decomposition.}

\item{tukey}{Logical. If \code{TRUE}, Tukey's transformation
is applied. If \code{FALSE}, a Box-Cox style transformation
is used.}

\item{base}{Numeric. The base for the logarithm if a logarithmic
transformation (\code{p=0}) is applied. Defaults to \code{exp(1)}
(natural logarithm).}
}
\value{
A list of class \code{"eda_mean_sweep"} with the following
components:
\describe{
\item{global}{The \bold{common} or \bold{global mean}}
\item{response}{The name of the response variable used in the analysis.}
\item{effects}{A named list of \bold{main and interaction effects}. Each element
is a named vector of centered effects, representing the deviations from
the adjusted mean attributable to that factor or interaction.}
\item{residuals}{A numeric vector of \bold{residuals} after the global mean
and all specified effects have been "swept out" (subtracted) from the
response variable.}
\item{long}{The original data frame with an added \code{residuals} column,
which can be useful for further exploratory plotting.}
}
}
\description{
Performs an exploratory decomposition of a numeric response variable into
additive components (global mean, main effects, interaction effects, and
residuals) using a sequential mean sweeping algorithm. It supports both
balanced and unbalanced designs, and can account for nested factor structures.
}
\details{
This function implements the value-splitting and sweeping procedure
central to Exploratory Data Analysis (EDA) of Analysis of Variance (ANOVA).
It systematically decomposes the response variable into additive
overlays, which when recombined, recover the original data.

The decomposition process is sequential: the global mean is first removed,
then main effects are calculated and subtracted, followed by interaction
effects up to \code{max_order}. Each effect is
calculated as the mean deviation from the previously swept \code{y}
for its respective levels, and then subtracted, leaving the remaining \code{y}
for subsequent effects or as residuals.

Nested factors are specifically handled by computing their effects
within each level of their parent factor ensuring appropriate variance
attribution.
\cr \cr
Unbalanced designs are supported by calculating group wise means
allowing for a robust decomposition even when cell counts are unequal.

\strong{Important consideration for factor order}: The order in which factors are
specified can significantly affect the decomposition, particularly when factors
are correlated or nested. Factors listed earlier in the arguments are "swept"
first and may absorb variation that might otherwise be attributed to factors listed
later in the arguments. To manage this, the \code{nesting} argument provides a
structured way to enforce a logical sweeping sequence, ensuring parent factors
are accounted for before their nested children.
}
\examples{
# A one-way analysis of chickwts. "weight" is the response and "feed" is
# the factor. First column passed to the function must be the response variable
# ("weight" in this example)
M0 <- eda_mean_sweep(chickwts, weight, feed)

# Global (overall) mean weight
M0$global

# Effect level values
M0$effects

# Compare residuals' spread to those of the effects
plot(M0, label = TRUE)

# A two-way analysis without replicates (i.e. one value per cell)
M0 <- eda_mean_sweep(inf_mort, perc, region, edu)
plot(M0, label = TRUE)

# A two -way analysis with replicates (i.e. multiple values per cell)
# Include 2-way interaction effects
M0 <- eda_mean_sweep(feav5_12, Weight, Level, Type, max_order = 2)
plot(M0, label = TRUE)

# A three-way analysis with nested factors. There are two embedded nests:
# Sample embedded under Run and Run embedded under Lot.
# Response variable is decomposed across ALL factors leaving 0 residuals
M0 <- eda_mean_sweep(feav5_14, Absorption, Lot, Run, Sample,
                     nesting = list(c("Lot", "Run"), c("Run","Sample")))
plot(M0, rotate = TRUE)

# A traditional ANOVA table can be generated from the eda_mean_sweep object
eda_anova_table(M0)

# A three-way analysis with one nested factor (State within Grouping)
# If there is just one nesting object, the nesting argument can be passed
# a c() object without the need of embedding it in a list() object
M0 <- eda_mean_sweep(feav1_5, votes, State, Year, Grouping,
                     nesting = c("Grouping", "State"))
plot(M0, rotate = TRUE)

# A three-way analysis with 2-way interactions
M0 <- eda_mean_sweep(feav6_8, Hard, Dentist, Method, Alloy, Temp, max_order = 2)
plot(M0, rotate = TRUE, order = FALSE) # Preserve factor order as entered in arguments
plot(M0, rotate = TRUE) # By default, factors are ordered by range

# A three-way analysis with 3-way interactions
M0 <- eda_mean_sweep(feav6_8, Hard, Dentist, Method, Alloy, Temp, max_order = 3)
plot(M0, rotate = TRUE)
}
\references{
Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (1991).
\emph{Fundamentals of Exploratory Analysis of Variance}. Wiley.
}
\seealso{
\code{\link{eda_anova_table}} for computing the ANOVA table (e.g., Sums of Squares,
Mean Squares, F-statistics) from the output of this function.
\code{\link{plot.eda_mean_sweep}} for visualizing the decomposed effects and residuals.
}
