% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_moments.R
\name{eda_moments}
\alias{eda_moments}
\title{Compute moments of a dataset}
\usage{
eda_moments(x)
}
\arguments{
\item{x}{A numeric vector representing the dataset.}
}
\value{
A named vector with the count, mean, variance, skewness,
and excess kurtosis of the data.
}
\description{
Computes the mean, variance, skewness, and excess kurtosis of a
given numeric vector.
}
