% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_pol.R
\name{eda_pol}
\alias{eda_pol}
\title{Median polish of two-way tables}
\usage{
eda_pol(
  x,
  row = NULL,
  col = NULL,
  val = NULL,
  stat = median,
  plot = TRUE,
  eps = 0.01,
  maxiter = 5,
  sort = FALSE,
  p = 1,
  tukey = FALSE,
  offset = 1e-05,
  col.quant = FALSE,
  colpal = "RdYlBu",
  adj.mar = TRUE,
  res.size = 1,
  row.size = 1,
  col.size = 1,
  res.txt = TRUE,
  label.txt = TRUE,
  base = exp(1),
  ...
)
}
\arguments{
\item{x}{A three column data frame.}

\item{row}{Name of column assigned to the row effect.}

\item{col}{Name of column assigned to the column effect.}

\item{val}{Name of column assigned to the response variable.}

\item{stat}{Polishing statistic (default is median).}

\item{plot}{Boolean determining if an output plot should be generated.}

\item{eps}{Convergence tolerance parameter.}

\item{maxiter}{Maximum number of iterations.}

\item{sort}{Boolean determining if the effects row/columns should be sorted.}

\item{p}{Re-expression power parameter.}

\item{tukey}{Boolean determining if Tukey's power transformation should used.
If FALSE, the Box-Cox transformation is adopted.}

\item{offset}{Offset to add to values if at leat one value is 0 and the power
is negative.}

\item{col.quant}{Boolean determining if a quantile classification scheme
should be used.}

\item{colpal}{Color palette to adopt.}

\item{adj.mar}{Boolean determining if margin width needs to accomodate labels.}

\item{res.size}{Size of residual values in plot \code{[0-1]}.}

\item{row.size}{Size of row effect values in plot \code{[0-1]}.}

\item{col.size}{Size of column effect values in plot \code{[0-1]}.}

\item{res.txt}{Boolean determining if values should be added to plot.}

\item{label.txt}{Boolean determining if margin and column labels should be
plotted.}

\item{base}{Base used with the log() function if \code{p} is \code{0}.}

\item{...}{Not used}
}
\value{
A list of class \code{eda_polish} with the following named
components:

\itemize{
\item \code{input} The median polish residuals with three columns: Column levels,
row levels and residual values.
\item \code{wide} The median polish residuals table in wide form.
\item \code{row}  Row effects table
\item \code{col}  Column effects table
\item \code{global}  Overall value (common value)
\item \code{iter}  Number of iterations before polish stabilizes.
\item \code{long}  Table of residuals, row effects, column effects and CV values in long form.
\item \code{power}  Transformation power applied to values prior to polishing.
\item \code{IQ_row}  Ratio between interquartile row effect values and 80th quantile of residuals.
\item \code{IQ_col}  Ratio between interquartile column effect values and 80th quantile of residuals.}
}
\description{
\code{eda_pol} Polishes two-way tables using median, means, or
any customizable functions.
}
\details{
The function performs a polish on a two way table. By default, it
applies a median polish, but other statistical summaries such
as the mean can be passed to the function via the \code{stat = } argument.
The function returns a list of row/column effects along with global and
residual values. It will also generate a colored table if \code{plot =
  TRUE}. \cr
For an N-way table with more than two factors, see \code{\link{eda_npol}}.
}
\examples{
M <- eda_pol(inf_mort, row = region, col = edu, val = perc)
}
\references{
\itemize{
\item Hoaglin, David C. and Mosteller, Frederick and Tukey, John W. (1985).
Exploring data tables, trends, and shapes.
\item Tukey, John W. 1977. Exploratory Data Analysis. Addison-Wesley
\item \href{../articles/polish.html}{Median polish article}}
}
