% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_qqpool.R
\name{eda_qqpool}
\alias{eda_qqpool}
\title{Multi-panel pooled values QQ plots}
\usage{
eda_qqpool(
  dat,
  x,
  fac,
  p = 1L,
  tukey = FALSE,
  base = exp(1),
  q.type = 5,
  resid = TRUE,
  stat = mean,
  show.par = FALSE,
  plot = TRUE,
  grey = 0.6,
  pch = 21,
  nrow = 1,
  p.col = "grey40",
  p.fill = "grey60",
  size = 1,
  text.size = 0.8,
  tail.pch = 21,
  tail.p.col = "grey70",
  tail.p.fill = NULL,
  tic.size = 0.8,
  alpha = 0.8,
  q = FALSE,
  tails = FALSE,
  med = FALSE,
  inner = 0.75,
  title = FALSE,
  ylab = NULL,
  ...
)
}
\arguments{
\item{dat}{Data frame.}

\item{x}{Continuous variable.}

\item{fac}{Categorical variable.}

\item{p}{Power transformation to apply to the continuous variable.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(\code{FALSE} adopts a Box-Cox transformation).}

\item{base}{Base used with the log() function if \code{p = 0}.}

\item{q.type}{An integer between 4 and 9 selecting one of the six quantile
algorithms (See \code{eda_fval} for a list of quantile algorithms).}

\item{resid}{Boolean determining if residuals should be plotted. Residuals
are computed using the \code{stat} parameter.}

\item{stat}{Statistic to use if residuals are to be computed. Currently
\code{mean} (default) or \code{median}.}

\item{show.par}{Boolean determining if power transformation should be
displayed in the plot.}

\item{plot}{Boolean determining if plot should be generated.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{pch}{Point symbol type.}

\item{nrow}{Define the number of rows for panel layout.}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point symbol size (0-1).}

\item{text.size}{Size for category text above the plot.}

\item{tail.pch}{Tail-end point symbol type (See \code{tails}).}

\item{tail.p.col}{Tail-end color for point symbol (See \code{tails}).}

\item{tail.p.fill}{Tail-end point fill color passed to \code{bg}
(Only used for \code{tail.pch} ranging from 21-25).}

\item{tic.size}{Size of tic labels (defaults to 0.8).}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}

\item{q}{Boolean determining if grey box highlighting the \code{inner}
region should be displayed.}

\item{tails}{Boolean determining if points outside of the \code{inner} region
should be symbolized differently. Tail-end points are symbolized via the
\code{tail.pch}, \code{tail.p.col} and \code{tail.p.fill} arguments.}

\item{med}{Boolean determining if median lines should be drawn.}

\item{inner}{Fraction of mid-values to highlight in \code{q} or \code{tails}.
Defaults to the inner 75 percent of values.}

\item{title}{Title to display. If set to \code{TRUE}, defaults to
\code{"Pooled residuals QQ plot"}. If set to \code{FALSE}, omits title from output.
Custom title can also be passed to this argument.}

\item{ylab}{Y and X axes labels.}

\item{...}{Not used}
}
\value{
Returns a dataframe with \code{x} (or its residuals if
\code{resid = TRUE}), \code{fac}, the f-value (by \code{fac} group), and
the matching pooled residuals.
}
\description{
\code{eda_qqpool} generates multi-panel pooled values QQ plots
for a continuous variable conditioned on a grouping variable. This plot is
most frequently used to compare residuals to pooled residuals.
}
\details{
The function will generate a multi-panel pooled values QQ plots.
The function defaults to a pooled residuals QQ plots which is the most common
use for this kind of plot.
}
\examples{

# Default output
singer <- lattice::singer
eda_qqpool(singer, height, voice.part, nrow = 2)

}
\references{
\itemize{
\item William S. Cleveland. Visualizing data. (1993)}
}
