% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_re.R
\name{eda_re}
\alias{eda_re}
\title{Re-expression function}
\usage{
eda_re(x, p = 0, tukey = FALSE, base = exp(1))
}
\arguments{
\item{x}{Vector}

\item{p}{Power transformation}

\item{tukey}{If set to TRUE, then adopt Tukey's power transformation, if
FALSE, adopt Box-Cox transformation}

\item{base}{Base used with the \code{log()} function}
}
\value{
Returns a vector of same length as input \code{x}
}
\description{
\code{eda_re} re-expresses a vector following the Tukey or
box-cox transformation.
}
\details{
The function is used to re-express data using one of two
transformation techniques: Box-Cox transformation (\code{tukey = FALSE})or
Tukey's power transformation (\code{tukey = TRUE}).
}
\examples{
x <- c(15, 28, 17, 73,  8, 83,  2)
eda_re(x, p=-1/3)
}
