% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_sim.R
\name{eda_sim}
\alias{eda_sim}
\title{Generate Data with Specific Skewness and Kurtosis}
\usage{
eda_sim(n, skewness = 0, kurtosis = 0, ...)
}
\arguments{
\item{n}{Integer. Number of data points to generate.}

\item{skewness}{Numeric. Desired skewness of the generated data.
Default is 0 (symmetric distribution).}

\item{kurtosis}{Numeric. Desired kurtosis of the generated data.
Default is 0 (normal distribution).}

\item{...}{Not used.}
}
\value{
A numeric vector of length \code{n} containing the generated data.
}
\description{
This function generates a dataset with a specified skewness and
 kurtosis using Fleishman's polynomial transformation.
}
\details{
The function uses Fleishman's polynomial transformation of the form:
\deqn{Y = a + bX + cX^2 + dX^3}
where \code{a}, \code{b}, \code{c}, and \code{d} are coefficients determined
to approximate the specified skewness and kurtosis. The coefficients are
solved using a numerical optimization approach based on minimizing the
residuals of Fleishman's equations.
}
\examples{

 # A normal distribution
 x <- eda_sim(5000, 0, 0)
 hist(x)

 # A right-skewed distribution
 x <- eda_sim(5000, 2, 0)
 hist(x)

 # A left-skewed distribution
 x <- eda_sim(5000, -2, 0)
 hist(x)

 # A uniform distribution
 # Note that this is unbounded which may result in outliers
 set.seed(21)
 x <- eda_sim(5000, 0, -10)
 hist(x, breaks = 20)

 # A "peaky" distribution
  set.seed(12)
  x <- eda_sim(5000, 0, 10)
  hist(x, breaks = 20)

  # Distributions can be modified so as to share the same mean and
  # variances. For example, to have x1 (a right skewed distribution) share
  # the  same mean and variance as x2 (a left skewed distribution), type:
  set.seed(432)
  x1 <- eda_sim(10000, 0.5, 0)
  x2 <- eda_sim(10000, -0.5, 0)

  x1_u <- mean(x1)
  x1_sd <- sd(x1)
  x2_sd <- sqrt(var(x2))
  x1b <- (x1 - x1_u) / x1_sd * x2_sd + mean(x2)
  mean(x1b)
  mean(x2)
  var(x1b)
  var(x2)

eda_dens(x1b, x2)
}
\references{
\itemize{
 \item Fleishman, A. I. (1978). A method for simulating non-normal
 distributions. Psychometrika, 43, 521–532.}
}
