% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_sim_defunct.R
\name{eda_sim_defunct}
\alias{eda_sim_defunct}
\title{Generate Data with Specific Skewness and Kurtosis}
\usage{
eda_sim_defunct(n, skewness = 0, kurtosis = 0, ...)
}
\arguments{
\item{n}{Integer. Number of data points to generate.}

\item{skewness}{Numeric. Desired skewness of the generated data.
Default is 0 (symmetric distribution).}

\item{kurtosis}{Numeric. Desired excess kurtosis of the generated data.
Default is 0 for a Normal distribution.}

\item{...}{Not used.}
}
\value{
A numeric vector of length \code{n} containing the generated data.
}
\description{
This function generates a dataset with a specified skewness and
 excess kurtosis using Fleishman's polynomial transformation.
}
\details{
The function uses Fleishman's polynomial transformation of the form:
\deqn{Y = a + bX + cX^2 + dX^3}
where \code{a}, \code{b}, \code{c}, and \code{d} are coefficients determined
to approximate the specified skewness and excess kurtosis. An excess kurtosis
is defined as the kurtosis of a Normal distribution (k=3) minus 3. Hence, an
excess kurtosis of 0 is that of a Normal distribution. The coefficients are
solved using a numerical optimization approach based on minimizing the
residuals of Fleishman's equations. The simulated values have a mean
of 0, but the variance can range from 1 for normal distributions to a variance
greater than 1 for non-normal distributions.
}
\examples{

 # A normal distribution
 x <- eda_sim(5000, 0, 0)
 hist(x)

 # A right-skewed distribution
 x <- eda_sim(5000, 2, 0)
 hist(x)

 # A left-skewed distribution
 x <- eda_sim(5000, -2, 0)
 hist(x)

 # A uniform distribution
 # Note that this is unbounded which may result in outliers
 set.seed(21)
 x <- eda_sim(5000, 0, -10)
 hist(x, breaks = 20)

 # A "peaky" distribution
  set.seed(12)
  x <- eda_sim(5000, 0, 10)
  hist(x, breaks = 20)

}
\references{
\itemize{
 \item Fleishman, A. I. (1978). A method for simulating non-normal
 distributions. Psychometrika, 43, 521–532.}
}
