% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_simdata.R
\name{eda_simdata}
\alias{eda_simdata}
\title{Simulate a given dataset using fleishman transformation}
\usage{
eda_simdata(x, n)
}
\arguments{
\item{x}{A numeric vector representing the dataset to match moments.}

\item{n}{An integer specifying the number of simulated data points to generate.}
}
\value{
A numeric vector of simulated data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Generates simulated data with the same mean, standard deviation,
skewness, and kurtosis as the input data.
}
\examples{
set.seed(4321)
nile  <- as.vector(Nile)
simnile <- eda_simdata(nile, 1000)
eda_qq(nile, simnile)
}
\references{
\itemize{
\item Fleishman, A. I. (1978). A method for simulating non-normal
distributions. Psychometrika, 43, 521–532.
\item Wicklin, R. (2013). Simulating Data with SAS (Appendix D: Functions
for Simulating Data by Using Fleishman’s Transformation). Cary, NC: SAS
Institute Inc. Retrieved from  https://tinyurl.com/4tustnph }
}
\seealso{
\itemize{
\item \code{\link[tukeyedar]{eda_sim}} for simulating a distribution}
}
