% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_sym.R
\name{eda_sym}
\alias{eda_sym}
\title{Symmetry QQ plot}
\usage{
eda_sym(
  x,
  p = 1L,
  tukey = FALSE,
  base = exp(1),
  q.type = 5,
  show.par = TRUE,
  grey = 0.6,
  pch = 21,
  p.col = "grey50",
  p.fill = "grey80",
  size = 0.8,
  alpha = 0.8,
  inner = 0.75,
  q = TRUE,
  qcol = rgb(0, 0, 0, 0.05),
  tails = FALSE,
  tail.pch = 21,
  tail.p.col = "grey70",
  tail.p.fill = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  t.size = 1.2,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{Vector of sample}

\item{p}{Power transformation to apply to \code{x}.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(\code{FALSE} adopts a Box-Cox transformation).}

\item{base}{Base used with the \code{log()} function if \code{px} or \code{py}
is \code{0}.}

\item{q.type}{An integer between 1 and 9 selecting one of the nine quantile
algorithms used to generate inner shaded region. (See \code{quantile}tile
function).}

\item{show.par}{Boolean determining if power parameter should be displayed.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{pch}{Point symbol type.}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point size (0-1)}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}

\item{inner}{Fraction of the input data considered as "mid values". Defaults to
75\%. Used  to define shaded region boundaries, \code{q}, or to identify
which of the tail-end points are to be symbolized differently, \code{tails}.}

\item{q}{Boolean determining if \code{inner} data region should be shaded.}

\item{qcol}{Fill color of inner quantile box.}

\item{tails}{Boolean determining if points outside of the \code{inner} region
should be symbolized differently. Tail-end points are symbolized via the
\code{tail.pch},  \code{tail.p.col} and \code{tail.p.fill} arguments.}

\item{tail.pch}{Tail-end point symbol type (See \code{tails}).}

\item{tail.p.col}{Tail-end color for point symbol (See \code{tails}).}

\item{tail.p.fill}{Tail-end point fill color passed to \code{bg}
(Only used for \code{tail.pch} ranging from 21-25).}

\item{xlab}{X label for output plot. Ignored if \code{x} is a dataframe.}

\item{ylab}{Y label for output plot. Ignored if \code{x} is a dataframe.}

\item{title}{Title to add to plot.}

\item{t.size}{Title size.}

\item{plot}{Boolean determining if plot should be generated.}

\item{...}{Not used}
}
\value{
Returns a dataframe of upper and lower halves
}
\description{
\code{eda_sym} Generates a symmetry QQ plot.
}
\details{
Generates a symmetry quantile plot that compares the lower half
of the sorted data to its upper half. If the distribution is perfectly
symmetrical, the points will fall on the line.
}
\examples{

 singer <- lattice::singer
 tenor1 <- subset(singer, voice.part == "Tenor 1", select = height, drop = TRUE )

 # Default plot
 eda_sym(tenor1)

 # To remove inner region grey box set q to FALSE
 eda_sym(tenor1, q = FALSE)

 # You can also choose to have the points outside of the inner region
 # symbolized differently
 eda_sym(tenor1, q = FALSE, tails = TRUE, tail.pch = 3)
}
\references{
\itemize{
\item John M. Chambers, William S. Cleveland, Beat Kleiner, Paul A. Tukey.
Graphical Methods for Data Analysis (1983)
\item \href{../articles/symqq.html}{Symmetry quantile plot article}}
}
