% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_theopan.R
\name{eda_theopan}
\alias{eda_theopan}
\title{Multi-panel theoretical QQ plots}
\usage{
eda_theopan(
  dat,
  x,
  fac,
  p = 1L,
  tukey = FALSE,
  base = exp(1),
  q.type = 5,
  dist = "norm",
  dist.l = list(),
  ylim = NULL,
  resid = FALSE,
  stat = mean,
  show.par = FALSE,
  plot = TRUE,
  grey = 0.6,
  pch = 21,
  nrow = 1,
  p.col = "grey40",
  p.fill = "grey60",
  size = 1,
  text.size = 0.8,
  tail.pch = 21,
  tail.p.col = "grey70",
  tail.p.fill = NULL,
  tic.size = 0.7,
  alpha = 0.8,
  q = FALSE,
  tails = FALSE,
  med = FALSE,
  inner = 0.75,
  iqr = TRUE,
  title = FALSE,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{dat}{Data frame.}

\item{x}{Continuous variable.}

\item{fac}{Categorical variable.}

\item{p}{Power transformation to apply to the continuous variable.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(\code{FALSE} adopts a Box-Cox transformation).}

\item{base}{Base used with the log() function if \code{p = 0}.}

\item{q.type}{An integer between 4 and 9 selecting one of the nine quantile
algorithms. (See \code{\link[tukeyedar]{eda_fval}} for a list of quantile
algorithms).}

\item{dist}{Theoretical distribution to use. Defaults to Normal distribution.}

\item{dist.l}{List of parameters passed to the distribution quantile function.}

\item{ylim}{Y axes limits.}

\item{resid}{Boolean determining if residuals should be plotted. Residuals
are computed using the \code{stat} parameter.}

\item{stat}{Statistic to use if residuals are to be computed. Currently
\code{mean} (default) or \code{median}.}

\item{show.par}{Boolean determining if power transformation should be
displayed in the plot.}

\item{plot}{Boolean determining if plot should be generated.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{pch}{Point symbol type.}

\item{nrow}{Define the number of rows for panel layout.}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point symbol size (0-1).}

\item{text.size}{Size for category text above the plot.}

\item{tail.pch}{Tail-end point symbol type (See \code{tails}).}

\item{tail.p.col}{Tail-end color for point symbol (See \code{tails}).}

\item{tail.p.fill}{Tail-end point fill color passed to \code{bg}
(Only used for \code{tail.pch} ranging from 21-25).}

\item{tic.size}{Size of tic labels (defaults to 0.8).}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}

\item{q}{Boolean determining if grey box highlighting the \code{inner}
region should be displayed.}

\item{tails}{Boolean determining if points outside of the \code{inner} region
should be symbolized differently. Tail-end points are symbolized via the
\code{tail.pch}, \code{tail.p.col} and \code{tail.p.fill} arguments.}

\item{med}{Boolean determining if median lines should be drawn.}

\item{inner}{Fraction of mid-values to highlight in \code{q} or \code{tails}.
Defaults to the inner 75 percent of values.}

\item{iqr}{Boolean determining if an IQR line should be fitted to the points.}

\item{title}{Title to display. If set to \code{TRUE}, defaults to
theoretical distribution type. If set to \code{FALSE}, omits title
from output. Custom title can also be passed to this argument.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{...}{Not used}
}
\value{
Returns a list with the following components:

\itemize{
\item \code{data}: List with input \code{x} and \code{y} values for each
group. May be interpolated to smallest quantile batch if batch sizes
don't match. Values will reflect power transformation defined in \code{p}}.
}
\description{
\code{eda_theopan} generates a multi-panel theoretical QQ plot
for a continuous variable conditioned on a grouping variable.
}
\details{
The function will generate a multi-panel theoretical QQ plot.
Currently, only the Normal QQ plot (\code{dist="norm"}), exponential
QQ plot (\code{dist="exp"}), uniform QQ plot (\code{dist="unif"}),
gamma QQ plot (\code{dist="gamma"}), chi-squared QQ plot
(\code{dist="chisq"}), and the Weibull QQ plot (\code{dist="weibull"}) are
currently supported. By default, the Normal QQ plot maps the unit Normal
quantiles to the x-axis (i.e. centered on a mean of 0 and standard deviation
of 1 unit).
}
\examples{

# Default output
singer <- lattice::singer
eda_theopan(singer, height, voice.part)

# Split into two rows
eda_theopan(singer, height, voice.part, nrow = 2, title = TRUE)

# Compare to a uniform distribution
eda_theopan(singer, height, voice.part, nrow = 2, dist = "unif")

# A uniform QQ plot is analogous to a Q(f) plot
eda_theopan(singer, height, voice.part, nrow = 2, dist = "unif",
              iqr = FALSE, xlab = "f-value")

# Normal QQ plots of Waterville daily averages. Mean monthly values are
# subtracted from the data to recenter all batches around 0.  Color and point
# symbols are used to emphasize the inner core of the data (here set to the
# inner 80\% of values)
wat <- tukeyedar::wat05
wat$month <- factor(format(wat$date,"\%b"), levels = month.abb)
eda_theopan(wat,avg, month, resid = TRUE, nrow = 3, inner = 0.8 ,
                    tails = TRUE, tail.pch = 3, p.fill = "coral")
}
\references{
\itemize{
\item William S. Cleveland. Visualizing data. (1993)}
}
\seealso{
\itemize{
\item \code{\link[tukeyedar]{eda_fval}} for computing f-values
\item \code{\link[tukeyedar]{eda_theo}} for a single theoretical QQ plot}
}
