% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{eda_trim}
\alias{eda_trim}
\alias{eda_trim_df}
\alias{eda_ltrim}
\alias{eda_ltrim_df}
\alias{eda_rtrim}
\alias{eda_rtrim_df}
\title{Trims vector and dataframe objects}
\usage{
eda_trim(x, prop = 0.05, num = 0)

eda_trim_df(dat, x, prop = 0.05, num = 0)

eda_ltrim(x, prop = 0.05, num = 0)

eda_ltrim_df(dat, x, prop = 0.05, num = 0)

eda_rtrim(x, prop = 0.05, num = 0)

eda_rtrim_df(dat, x, prop = 0.05, num = 0)
}
\arguments{
\item{x}{Vector of values (if trimming a vector) or the column whose
values are used to trim a dataframe (applies to \code{*_df}
functions only)}

\item{prop}{Fraction of values to trim}

\item{num}{Number of values to trim}

\item{dat}{Dataframe (applies to \code{*_df} functions only)}
}
\description{
Removes records from either tail-ends of a sorted dataset. Trimming can be
 performed by number of records (specify the \code{num =} option) or by
 quantiles (specify the \code{prop=} option). \cr\cr
 \code{eda_trim} Trims a vector \cr
 \code{eda_trim_df} Trims a data frame \cr
 \code{eda_ltrim} Left-trims a vector \cr
 \code{eda_rtrim} Right-trims a vector \cr
 \code{eda_ltrim_df} Left-trims a dataframe \cr
 \code{eda_rtrim_df} Right-trims a dataframe \cr
}
\details{
\itemize{
 \item The input dataset does not need to be sorted (sorting is performed in the
 functions).
 \item If \code{num} is set to zero, then the function will assume that the
 trimming is to be done by fraction (defined by the \code{prop} parameter).
 \item If \code{eda_trim} or \code{eda_trim_df} functions are called, the
 \code{num} and \code{prop} values apply to each tail. For example, if
 \code{num = 5} then the 5 smallest AND 5 largest values are removed from the
 data.
 \item \code{NA} values must be stripped from the input vector or
 column elements before running the trim functions.
 \item Elements are returned sorted on the trimmed elements.
 }
}
\examples{

# Trim a vector by 10\% (i.e. 10\% of the smallest and 10\% of the largest
# values)
eda_trim( mtcars[,1], prop=0.1)

# Trim a data frame by 10\% using the mpg column(i.e. 10\% of the smallest
# and 10\% of the largest mpg values)
eda_trim_df( mtcars, mpg, prop=0.1)
}
\keyword{trim}
