% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edtts_data.R
\docType{data}
\name{edtts2.12}
\alias{edtts2.12}
\title{Confidence in American institutions survey data}
\format{
A data frame with the following variables:
\describe{
\item{institution}{American institutions}
\item{year}{Survey year}
\item{perc}{Percentage of respondents indicating "a great deal" or
quite a lot" of confidence in American institutions}
}
}
\usage{
data(edtts2.12)
}
\description{
Table 2-12 in \emph{Exploring Data tables, Trends, and Shapes} listing the
percentage of respondents indicating "a great deal" or "quite a lot" of
confidence in American institutions between the years 1973 and 1980.
}
\references{
\itemize{
\item{Hoaglin, David C. and Mosteller, Frederick and Tukey, John W. (1985).
Exploring data tables, trends, and shapes. Wiley.}
}
}
\keyword{datasets}
