% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edtts_data.R
\docType{data}
\name{edtts4.2}
\alias{edtts4.2}
\title{Weight Gain in Pigs Data}
\format{
A data frame with the following variables:
\describe{
\item{Methionine}{Methionine offered at 3 levels: 0\%, 0.25\% amd 0.50\%}
\item{Protein}{Soybean meal protein offered at 2 levels: 12\% and 14\%}
\item{Lysine}{Lysine offered at 4 levels: 0\%, 0.05\%, 0.1\% and 0.15\%}
\item{WeightGain}{Treatment totals of average daily weight gains
(measured in units of .01 lb) for the combined
results for two pigs}
}
}
\usage{
data(edtts4.2)
}
\description{
Table 4-2 in \emph{Exploring Data tables, Trends, and Shapes} is
derived from an experiment conducted at the Iowa Agricultural Experiment
Station. It summarizes the weight gain of male pigs across various combinations
of three dietary supplements
}
\references{
\itemize{
\item{Hoaglin, David C. and Mosteller, Frederick and Tukey, John W. (1985).
Exploring data tables, trends, and shapes. Wiley.}
}
}
\keyword{datasets}
