% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edtts_data.R
\docType{data}
\name{edtts4.25}
\alias{edtts4.25}
\title{Dataset that is perfectly additive in the square root scale}
\format{
A data frame with the following variables:
\describe{
\item{A}{Effect A}
\item{B}{Effect B}
\item{C}{Effect C}
\item{Value}{Response variable}
}
}
\usage{
data(edtts4.25)
}
\description{
Table 4-25 of \emph{Exploring Data tables, Trends, and Shapes} is a synthetic
dataset of a three-way table. The key characteristic of this data is that if
one takes the square root of each value in the table, the resulting
values would perfectly fit a simple additive model. This means that, in the
square root scale, there are no interaction effects between the factors.
}
\details{
The data is constructed using specific values
\itemize{
\item{grand mean:} \eqn{\mu = 20}
\item{factor A:} \eqn{(\alpha_1, \alpha_2, \alpha_3) = (-1, 0, 3)}
\item{factor B:} \eqn{(\beta_1, \beta_2, \beta_3) = (-1, 0, 1)}
\item{factor C:} \eqn{(\gamma_1, \gamma_2, \gamma_3) = (-1, 0, 2)}
}
}
\examples{
# Median polish of raw values
M0 <- eda_npol(edtts4.25, Value, A, B, C)

# There is evidence of strong interaction as shown in this plot
plot(M0, plot = "diagnostic")

# Taking the square root eliminates interaction effects
# Note that this may throw a warning if loess option is
# set to TRUE
M1 <- eda_npol(edtts4.25, Value, A, B, C, p = 0.5)
plot(M1, plot = "diagnostic", loe = FALSE)
}
\references{
\itemize{
\item{Hoaglin, David C. and Mosteller, Frederick and Tukey, John W. (1985).
Exploring data tables, trends, and shapes. Wiley.}
}
}
\keyword{datasets}
