% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feav_data.R
\docType{data}
\name{feav5_12}
\alias{feav5_12}
\title{Weight Gain in Rats on Different Diets}
\format{
A \code{data.frame} with 60 rows and the following columns:

\describe{
\item{Level}{A character indicating the amount of protein.
Either \code{High} or \code{Low}. }
\item{Type}{A character indicating the specific diet administered
to each group of rats.}
\item{Weight}{An integer vector representing the weight gain,
in grams, for each rat. This is the response variable.}
}
}
\source{
Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (1991).
\emph{Fundamentals of Exploratory Analysis of Variance}. Wiley.
}
\usage{
feav5_12
}
\description{
This dataset contains the weight gain (in grams) of
rats subjected to different dietary treatments. Data are pulled from table
5-12 of the referenced source.
\cr \cr
This dataset serves as an example of a two-way table with (balanced)
replicates (10 replicates per \code{Level} and \code{Type} combinations).
}
\examples{
# Include second level interactions
M0 <- eda_mean_sweep(feav5_12, Weight, Level, Type, max_order = 2)
plot(M0, label = TRUE)

}
\keyword{datasets}
