% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feav_data.R
\docType{data}
\name{feav5_14}
\alias{feav5_14}
\title{ELISA Test Absorbance for HIV Positive Controls}
\format{
A \code{data.frame} with 75 rows and the following columns:

\describe{
\item{Lot}{A character indicating the specific production lot
of the ELISA test \code{A, B, C, D, or E}.}
\item{Run}{A character  representing the individual test runs.
\code{Run} is nested within \code{Lot}, meaning \code{Run 1} for
\code{Lot A} is distinct from \code{Run 1} for \code{Lot B}.}
\item{Sample}{A character specifying the sample. There are
three replicate samples per run.}
\item{Absorption}{A numeric vector representing the light
absorbance readings (in arbitrary units) for the positive control
samples. This is the response variable.}
}
}
\source{
Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (1991).
\emph{Fundamentals of Exploratory Analysis of Variance}. Wiley.
}
\usage{
feav5_14
}
\description{
This dataset contains measurements of light absorbance for
positive control samples from an ELISA (Enzyme-Linked Immunosorbent Assay)
test for HIV. Data are pulled from table 5-14 of the referenced
source.
\cr \cr
The dataset is an example of a two-factor experimental design with nested
factors. The experiment involved five different production lots. For
each lot, five individual runs of the test were performed. Within each run,
absorbance readings were recorded for three positive control samples. As such
the \code{Sample} factor is nested under the \code{Run} factor which is itself
nested under the \code{lot} factor.
}
\examples{
# Partition response variable across ALL factors. Residuals should be 0.
M0 <- eda_mean_sweep(feav5_14, Absorption, Lot, Run, Sample,
                     nesting = list(c("Lot", "Run"), c("Run","Sample")))
plot(M0, rotate = TRUE)

}
\keyword{datasets}
