% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feav_data.R
\docType{data}
\name{feav6_1}
\alias{feav6_1}
\title{Simulated Tumor Measurements by Oncologists}
\format{
A \code{data.frame} with 78 rows and the following columns:

\describe{
\item{Oncologist}{A character uniquely identifying each
oncologist who provided a cross-sectional area estimate.}
\item{Material}{A character indicating the material of the
simulated tumor model (\code{Cork or Rubber}).}
\item{Form}{A character representing the shape of the simulated
tumor (\code{Small, Oblong or Large}).}
\item{Area}{A numeric vector representing the average cross-sectional
area (in mm²) of the tumor as estimated by the oncologist. This
is the response variable.}
}
}
\source{
Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (1991).
\emph{Fundamentals of Exploratory Analysis of Variance}. Wiley.
}
\usage{
feav6_1
}
\description{
This dataset contains measurements of cross-sectional area (in
mm²) of simulated solid tumors. These measurements were estimated by 13
different oncologists. Data are pulled from table 6-1 of the referenced
source.
\cr \cr
This dataset represents a three-factor experimental design with fully
crossed factors. The primary factors involved are the Oncologist (the
individual performing the estimate), the Material from which the tumor
model was made (Cork or Rubber), and the Form (or shape) of the tumor
(Small, Oblong, or Large).
}
\examples{
M0 <- eda_mean_sweep(feav6_1, Area, Oncologist, Material, Form, max_order = 2)
plot(M0, order = FALSE, rotate = TRUE)

}
\keyword{datasets}
