% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feav_data.R
\docType{data}
\name{feav6_8}
\alias{feav6_8}
\title{Diamond Pyramid Hardness Number of Dental Fillings}
\format{
A \code{data.frame} with 90 rows and the following columns:

\describe{
\item{Dentist}{A factor with 5 levels, each identifying a denstist.}
\item{Method}{A factor denoting the condensation method used.}
\item{Alloy}{A character indicating the type of gold alloy used.}
\item{Temp}{A character representing the sintering temperature.}
\item{Hard}{An integer vector representing the summed Diamond
Pyramid Hardness Number (DPHN) for each combination of factors.
This is the response variable.}
}
}
\source{
Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (1991).
\emph{Fundamentals of Exploratory Analysis of Variance}. Wiley.
}
\usage{
feav6_8
}
\description{
This dataset contains measurements of the Diamond Pyramid
Hardness Number (DPHN) of simulated dental fillings. The response variable
represents the sum of 10 individual DPHN readings for each sample.
Data are pulled from table 6-8 of the referenced source.
\cr \cr
The data were collected from a study involving five different
dentists. The dental fillings were made from two types of gold alloy
(\code{A1} and \code{AuCa}), sintered at three different temperatures
(\code{1500F}, \code{1600F}, \code{1700F}), and prepared using
three condensation methods (\code{1}, \code{2}, \code{3}).
}
\examples{
M0 <- eda_mean_sweep(feav6_8, Hard, Dentist, Method, Alloy, Temp, max_order = 3)
plot(M0, rotate = TRUE, order = FALSE)

}
\keyword{datasets}
