% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\docType{data}
\name{income}
\alias{income}
\title{Income dataset}
\format{
A data frame with 10 rows and the following variables:
\describe{
\item{Education}{Level of educational attainment
\itemize{
\item{\code{NoHS}}: Less than High School Graduate
\item{\code{HS}}: High School Graduate (Includes Equivalency)
\item{\code{AD}}:Some College or Associate’s Degree
\item{\code{Grad}}:Bachelor’s Degree or Greater
}}

\item{Sex}{Male or Female}
\item{Earnings}{Earnings in 2021 dollars}
}
}
\usage{
data(income)
}
\description{
Earnings by sex and educational attainment for 2021 (source: US Census Bureau)
}
\keyword{datasets}
