% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incrent22}
\alias{incrent22}
\title{American Community Survey 2022 Median household income and median rent by county}
\format{
A data frame with 3222 rows and 5 variables:
\describe{
\item{FIPS}{County/state FIPS code.}
\item{County}{County name}
\item{State}{State name.}
\item{Income}{Median Household Income in 2022 US Dollars.}
\item{Rent}{Median Gross Rent in 2022 US Dollars.}
...
}
}
\source{
https://www.census.gov/
}
\usage{
incrent22
}
\description{
American Community Survey 2022 Median Household Income (B19013001) and
Median Gross Rent (B25064001) by county.  Units are in 2022 inflation-Adjusted
Dollars.
}
\keyword{datasets}
