% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_mort.R
\docType{data}
\name{inf_mort}
\alias{inf_mort}
\title{Infant mortality dataset}
\format{
A data frame with the following variables:
\describe{
\item{region}{Regional effect}
\item{edu}{Father's education effect
\itemize{
\item{\code{ed8}: Less than High School Graduate}
\item{\code{ed12}: High School Graduate}
\item{\code{ed13to15}: Some College or Associate’s Degree}
\item{\code{ed16}: Bachelor’s Degree or Greater}
}}
\item{perc}{Numbers of infant deaths per 1000 live births}
}
}
\usage{
data(inf_mort)
}
\description{
Infant mortality dataset pulled from Table 6-2 of "Understanding
Robust and Exploratory Data Analysis".
}
\references{
\itemize{
\item{Emerson, John D., and David C. Hoaglin. (1983). Understanding
Robust and Exploratory Data Analysis. John Wiley & Sons.}
}
}
\keyword{datasets}
