% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neoplasms}
\alias{neoplasms}
\title{Breast cancer mortality vs. temperature}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{Temp}{Temperature in degrees Fahrenheit.}
  \item{Mortality}{Mortality rate presented as an index.}
  ...
}
}
\source{
Applications, Basics and Computing of Exploratory Data Analysis,
      P.F. Velleman and D.C. Hoaglin, 1981. (page 127)
}
\usage{
neoplasms
}
\description{
The data represent the relationship between mean annual temperature
and breast cancer mortality rate.
}
\keyword{datasets}
