% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nine_point}
\alias{nine_point}
\title{Andrew Siegel's pathological 9-point dataset}
\format{
A data frame with 9 rows and 2 variables:
\describe{
  \item{X}{X values}
  \item{Y}{Y values}
  ...
}
}
\source{
Robust regression using repeated medians, Andrew F. Siegel, Biometrika,
   vol 69, n 1, 1982.

Understanding robust and exploratory data analysis, by D.C. Hoaglin,
   F. Mosteller and J.W. Tukey. 1983 (page 139)
}
\usage{
nine_point
}
\description{
A synthetic dataset created to test the robustness of fitted lines.
Originally published by Andrew Siegel and later adapted in Hoaglin et al.'s
book.
}
\keyword{datasets}
