% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eda_rline.R
\name{plot.eda_rline}
\alias{plot.eda_rline}
\title{Plot eda_rline model}
\usage{
\method{plot}{eda_rline}(
  x,
  plot = "model",
  xlab = NULL,
  ylab = NULL,
  model = TRUE,
  pt3 = TRUE,
  seg = TRUE,
  fit = TRUE,
  fit.col = rgb(1, 0, 0),
  grey = 0.6,
  sd = FALSE,
  mean.l = FALSE,
  reg = FALSE,
  lm.col = rgb(0.7, 0.7, 1, 0.8),
  lm.lty = 2,
  loe = NULL,
  loe.col = rgb(0.7, 0.7, 1, 0.8),
  loe.lty = 2,
  ...
)
}
\arguments{
\item{x}{Object of class \code{eda_rline}.}

\item{plot}{Plot type. One of two: "model", "residuals".}

\item{xlab}{Custom x-axis label. Defaults to column name.}

\item{ylab}{Custom y-axis label. Defaults to column name.}

\item{model}{Logical; whether to show  the resulting model above
plot. Only applies to \code{type = "model"}.}

\item{pt3}{Logical; whether to show the 3-pt summaries.
Only applies to \code{type = "model"}.}

\item{seg}{Logical; whether line segments showing partitioning of x variables
should be displayed.}

\item{fit}{Logical; whether to show the resistant fitted line.}

\item{fit.col}{Resistant line color.}

\item{grey}{Numeric between 0–1; controls grayscale background elements (0 =
black, 1 = white).}

\item{sd}{Logical; whether to show ±1 SD lines.}

\item{mean.l}{Logical; whether to show mean reference lines.}

\item{reg}{Logical; whether to fit and display an OLS regression line.}

\item{lm.col}{OLS regression line color (only applicable if \code{reg = TRUE}.}

\item{lm.lty}{OLS regression line line type (only applicable if \code{reg = TRUE}.}

\item{loe}{Logical; whether to plot loess smooth line.}

\item{loe.col}{loess smooth line color (only applicable if \code{loe = TRUE}.}

\item{loe.lty}{loess smooth line type (only applicable if \code{loe = TRUE}.}

\item{...}{Arguments to be passed to \code{.eda_plot_xy}.}
}
\value{
Returns the margins (\code{mar}) used to generate the plot via the
\code{par()} function. This parameter can be set with a subsequent call to
\code{par()} if additional elements are to be added to the plot.
}
\description{
A plot method for lists of \code{eda_rline} class.
}
\details{
The function generates a scatter plot with the fitted model from an
\code{eda_rline} object.
}
\examples{
r.lm <- eda_rline(age_height, Months, Height)

plot(r.lm)
plot(r.lm, pt3 = FALSE)

# And an OLS regression line for comparison (defaults to blue dashed line)
plot(r.lm, reg = TRUE)

# Plot residuals vs dependence plot. A Loess smooth is fitted to the data.
plot(r.lm, plot = "residuals")
}
