% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyedar-package.R, R/tukeyedar.R
\docType{package}
\name{tukeyedar-package}
\alias{tukeyedar-package}
\title{tukeyedar: Tukey Inspired Exploratory Data Analysis Functions}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A set of Tukey inspired functions for use in exploring datasets in a robust manner.

This packages hosts a small set of Tukey inspired functions for use
in exploring datasets in a robust manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mgimond.github.io/tukeyedar/}
  \item Report bugs at \url{https://github.com/mgimond/tukeyedar/issues/}
}

}
\author{
\strong{Maintainer}: Manuel Gimond \email{mgimond@gmail.com} (\href{https://orcid.org/0000-0002-3687-2316}{ORCID})

Authors:
\itemize{
  \item Manuel Gimond \email{mgimond@gmail.com} (\href{https://orcid.org/0000-0002-3687-2316}{ORCID})
}

}
\keyword{internal}
