% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wat05}
\alias{wat05}
\title{Temperature normals for Waterville Maine (1991-2020)}
\format{
A data frame with 365 rows and 5 variables:
\describe{
\item{date}{Date centered on the 11991-2020 period. Note that the year is purely symbolic.}
\item{doy}{Day of year.}
\item{min}{Typical minimum temperature for the 1991-2020 period.}
\item{avg}{Typical average temperature for the 1991-2020 period.}
\item{max}{Typical maximum temperature for the 1991-2020 period.}
...
}
}
\source{
https://www.ncei.noaa.gov/
}
\usage{
wat05
}
\description{
NOAA/NCEI derived normal daily temperatures for the city of Waterville, Maine
(USA) for the 1991 to 2020 period. Units are in degrees Fahrenheit.
}
\keyword{datasets}
