% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wat95}
\alias{wat95}
\title{Legacy temperature normals for Waterville Maine (1981-2010)}
\format{
A data frame with 365 rows and 5 variables:
\describe{
\item{date}{Date centered on the 1981-2010 period. Note that the year is purely symbolic.}
\item{doy}{Day of year}
\item{min}{Typical minimum temperature for the 1981-2010 period.}
\item{avg}{Typical average temperature for the 1981-2010 period.}
\item{max}{Typical maximum temperature for the 1981-2010 period.}
...
}
}
\source{
https://www.ncei.noaa.gov/
}
\usage{
wat95
}
\description{
NOAA/NCEI derived normal daily temperatures for the city of Waterville, Maine
(USA) for the 1981 to 2010 period. Units are in degrees Fahrenheit.
}
\keyword{datasets}
