% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edtts_data.R
\docType{data}
\name{yarn}
\alias{yarn}
\title{yarn dataset}
\format{
A data frame with the following variables:
\describe{
\item{Length}{Length effect}
\item{Amplitude}{Amplitude effect}
\item{Load}{Load effect}
\item{Cycles}{Response variable}
}
}
\usage{
data(yarn)
}
\description{
Table 4-19 of \emph{Exploring data tables, trends, and shapes} shows data from a
textile experiment where specimens of worsted yarn were repeatedly loaded
until they broke. The table shows the resulting number of cycles to failure
for the yarn under different conditions related to the length of the specimen,
the amplitude of the loading, and the load applied.
}
\references{
\itemize{
\item{Hoaglin, David C. and Mosteller, Frederick and Tukey, John W. (1985).
Exploring data tables, trends, and shapes. Wiley.}
}
}
\keyword{datasets}
