% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_3pt.R
\name{eda_3pt}
\alias{eda_3pt}
\title{3-point summary plot}
\usage{
eda_3pt(
  dat,
  x,
  y,
  x.lab = NULL,
  y.lab = NULL,
  adj = -0.12,
  dir = TRUE,
  pch = 20,
  col = "grey40",
  ...
)
}
\arguments{
\item{dat}{Data frame}

\item{x}{Column name assigned the x axis}

\item{y}{Column name assigned the y axis}

\item{x.lab}{X label for output plot}

\item{y.lab}{Y label for output plot}

\item{adj}{Adjustment parameter for y label}

\item{dir}{Boolean indicating if suggested ladder of power direction should
be displayed}

\item{pch}{Plot point size as a fraction (can be larger than 1.0)}

\item{col}{Plot point color}

\item{...}{Other parameters passed to the graphics::plot function.}
}
\value{
Generates a plot and returns a list with the following named
  components:

\itemize{
  \item \code{hsrtio}: The ratio between both slopes. A value close to one
  suggests that no transformation is needed.
  \item \code{xmed}: The x-coordinate values for the three summary points.
  \item \code{ymed}: The y-coordinate values for the three summary points.}
}
\description{
\code{eda_3pt} splits the data into 3 groups (whose summary
locations are defined by their respective medians), and two half slopes
linking the groups. The function will return a scatter plot showing the
half-slopes in red solid lines. The solid grey slope linking both tail-end
groups shows the desired shape for both half-slopes. The goal is to have the
two halve slopes line up as closely as possible to the solid grey slope via
re-expression techniques when seeking a linear relationship between both
variables.

The function will also return the half-slopes ratio \code{hsrtio} and the
direction of re-expression for both X and Y values on the ladder of powers.
}
\details{
Computes the three-point summary originally defined in Tukey's EDA
book (see reference).
}
\examples{

hsratio <- eda_3pt(cars, speed, dist)
hsratio <- eda_3pt(cars, speed, dist^(1/3), y.lab=expression("Dist"^{1/3}), adj=-0.1)
}
\references{
\itemize{
   \item Velleman, P. F., and D. C. Hoaglin. 1981. Applications, Basics and Computing of Exploratory Data Analysis. Boston: Duxbury Press.
   \item D. C. Hoaglin, F. Mosteller, and J. W. Tukey. 1983. Understanding Robust and Exploratory Data Analysis. Wiley.
   \item Tukey, John W. 1977. Exploratory Data Analysis. Addison-Wesley.}
}
