% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_bipow.R
\name{eda_bipow}
\alias{eda_bipow}
\title{Ladder of powers transformation on bivariate data with three-point summary plot}
\usage{
eda_bipow(dat, x, y, p = c(-1, 0, 0.5, 1, 2), tukey = FALSE, ...)
}
\arguments{
\item{dat}{Dataframe.}

\item{x}{Variable assigned to the x axis.}

\item{y}{Variable assigned to the y axis.}

\item{p}{Vector of powers.}

\item{tukey}{If set to TRUE, then adopt Tukey's power transformation. If FALSE,
adopt Box-Cox transformation.}

\item{...}{Other parameters passed to the graphics::plot function.}
}
\value{
{No return value}
}
\description{
Re-expresses a vector on the ladder of powers.
Requires \code{eda_3pt()} function.
}
\details{
Generates a matrix of scatter plots and boxplots of various
re-expressions of both x and y values. The 3-point summary and
associated half-slopes are also plotted (this function makes use
of the eda_3pt function). The values are re-expressed using either
the Tukey power transformation or the Box-Cox transformation (the default).
See \code{eda_re} for more information on these transformation techniques.
Axes labels are omitted to reduce plot clutter.
}
\examples{

data(cars)
# Example 1
eda_bipow(dat = cars, x = speed, y = dist)
# Custom powers
eda_bipow(dat = cars, x = speed, y = dist, p = c(-1, -0.5, 0, 0.5, 1))
# Adopt Tukey transformation
eda_bipow(dat = cars, x = speed, y = dist, tukey = TRUE, p = c(-1, -0.5, 0, 0.5, 1))

}
\references{
\itemize{
\item Tukey, John W. 1977. Exploratory Data Analysis. Addison-Wesley.}
}
