% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_bipow.R
\name{eda_bipow}
\alias{eda_bipow}
\title{Ladder of powers transformation on bivariate data with three-point summary plot}
\usage{
eda_bipow(x, y, dat, p = c(3, 2, 1, 0.5, 0), tukey = TRUE, ...)
}
\arguments{
\item{x}{column name assigned the x axis}

\item{y}{column name assigned the y axis}

\item{dat}{data frame}

\item{p}{vector of powers}

\item{tukey}{if set to TRUE then adopt Tukey's power transformation, if FALSE,
adopt Box-Cox transformation technique}

\item{...}{other parameters passed to the graphics::plot function.}
}
\description{
To re-express a vector on the ladder of powers.
 Requires eda_3pt() function.
}
\details{
References
  Applications, Basics and Computing of Exploratory Data Analysis,
   by P.F. Velleman and D.C. Hoaglin
  Understanding robust and exploratory data analysis, by D.C. Hoaglin,
   F. Mosteller and J.W. Tukey
  Exploratory Data Analysis, by John Tukey
}
