% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_boxls.R
\name{eda_boxls}
\alias{eda_boxls}
\title{Create boxplots equalized by level and spread}
\usage{
eda_boxls(dat, x, fac, outlier = TRUE, out.txt, type = "l", horiz = FALSE,
  outliers = TRUE)
}
\arguments{
\item{dat}{Data frame name}

\item{x}{Column name assigned to the values}

\item{fac}{Column name assigned to the factor the values are to be
conditioned on}

\item{outlier}{Boolean indicating if outliers should be plotted}

\item{out.txt}{Column whose values are to be used to label outliers}

\item{type}{Plot type. "none" = no equalization ; "l" = equalize by level;
"ls" = equalize by both level and spread}

\item{horiz}{plot horizontally (TRUE) or vertically (FALSE)}

\item{outliers}{plot outliers (TRUE) or not (FALSE)}
}
\description{
\code{eda_boxls} creates boxplots conditioned on one variable while providing the
option to equalize levels and/or spreads.
}
\examples{

# A basic boxplot (no equalization)
eda_boxls(mtcars,mpg, cyl, type="none", out.txt=mpg )

# Boxplots equalized by level
eda_boxls(mtcars,mpg, cyl, type="l", out.txt=mpg )

# Boxplots equalized by level and spread
eda_boxls(mtcars,mpg, cyl, type="ls", out.txt=mpg )

# Hide outlier
eda_boxls(mtcars,mpg, cyl, type="ls", out.txt=mpg , outlier=FALSE)

}
\keyword{boxplot}
\keyword{level}
\keyword{spread}
