% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_boxls.R
\name{eda_boxls}
\alias{eda_boxls}
\title{Create boxplots equalized by level and spread}
\usage{
eda_boxls(
  dat,
  x,
  fac,
  p = 1,
  tukey = FALSE,
  outlier = TRUE,
  out.txt,
  type = "none",
  horiz = FALSE,
  outliers = TRUE,
  grey = 0.6
)
}
\arguments{
\item{dat}{Data frame}

\item{x}{Column name assigned to the values}

\item{fac}{Column name assigned to the factor the values are to be
conditioned on}

\item{p}{Power transformation to apply to variable}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation)}

\item{outlier}{Boolean indicating if outliers should be plotted}

\item{out.txt}{Column whose values are to be used to label outliers}

\item{type}{Plot type. "none" = no equalization ; "l" = equalize by level;
"ls" = equalize by both level and spread}

\item{horiz}{plot horizontally (TRUE) or vertically (FALSE)}

\item{outliers}{plot outliers (TRUE) or not (FALSE)}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black)}
}
\value{
{No return value}
}
\description{
\code{eda_boxls} creates boxplots conditioned on one variable while providing the
option to equalize levels and/or spreads.
}
\examples{

# A basic boxplot (no equalization)
eda_boxls(mtcars,mpg, cyl, type="none", out.txt=mpg )

# Boxplots equalized by level
eda_boxls(mtcars,mpg, cyl, type="l", out.txt=mpg )

# Boxplots equalized by level and spread
eda_boxls(mtcars,mpg, cyl, type="ls", out.txt=mpg )

# Hide outlier
eda_boxls(mtcars,mpg, cyl, type="ls", out.txt=mpg , outlier=FALSE)

# For long factor level names, flip plot
eda_boxls(iris, Sepal.Length, Species, out.txt=Sepal.Length , horiz = TRUE)
}
