% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_dens.R
\name{eda_dens}
\alias{eda_dens}
\title{Overlapping density distributions for two variables}
\usage{
eda_dens(
  x,
  y,
  p = 1L,
  q.type = 5,
  tukey = FALSE,
  fx = NULL,
  fy = NULL,
  grey = 0.7,
  col = "red",
  size = 0.8,
  alpha = 0.4,
  xlab = NULL,
  ylab = NULL,
  legend = TRUE,
  ...
)
}
\arguments{
\item{x}{Vector for first variable.}

\item{y}{Vector for second variable.}

\item{p}{Power transformation to apply to both sets of values.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation).}

\item{fx}{Formula to apply to x variable. This is computed after any
transformation is applied to the x variable.}

\item{fy}{Formula to apply to y variable. This is computed after any
transformation is applied to the y variable.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{col}{Fill color for second density distribution.}

\item{size}{Point size (0-1).}

\item{alpha}{Fill transparency (0 = transparent, 1 = opaque). Only applicable
if \code{rgb()} is not used to define fill colors.}

\item{xlab}{X label for output plot.}

\item{ylab}{Y label for output plot.}

\item{legend}{Boolean determining if a legend should be added to the plot.}

\item{...}{Arguments passed to the \code{stats::density()} function.}
}
\value{
Does not return a value.
}
\description{
\code{eda_dens} generates overlapping density distributions for
  two variables.
}
\details{
This function will generate overlapping density plots with the first
  variable assigned a grey color and the second variable assigned the default
  red color.
}
\examples{

 # Accepting data as two separate vector objects
 set.seed(207)
 x <- rbeta(1000,2,8)
 y <- x * 1.5 + 0.1
 eda_dens(x, y)
}
