% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_fval.R
\name{eda_fval}
\alias{eda_fval}
\title{Calculate Quantile f-values}
\usage{
eda_fval(x, qtype = 5)
}
\arguments{
\item{x}{Vector whose f-values are to be computed.}

\item{qtype}{Quantile algorithm.}
}
\value{
A vector of \eqn{f-values} in the same order a vector \code{x}.
}
\description{
\code{eda_qq} Generates an empirical,  Normal,
 symmetry or Tukey mean-difference plot
}
\details{
For each value \eqn{i} in \code{x}, \code{eda_fval} computes the
cumulative probability associated with \eqn{i}. It provides a measure of the
position of a data point relative to the rest of the data scaled to the
range [0,1]. This is essentially the fraction, \eqn{f-value},  of the dataset
that lies below \eqn{i}. \cr
\cr
Different algorithms are used to compute the probability/f-value. This
function offers as option algorithm type \code{4} through \code{6} documented
in \code{stats::quantile}. The algorithms used are: \cr \cr
  \itemize{
    \item{\code{4}: \deqn{f = i / n,}}
    \item{\code{5}: \deqn{f = (i - 0.5) / n,}}
    \item{\code{6}: \deqn{f = i / (n + 1),}}
    \item{\code{7}: \deqn{f = (i  - 1) / (n - 1),}}
    \item{\code{8}: \deqn{f = (i - 1/3) / (n + 1/3),}}
    \item{\code{9}: \deqn{f = (i - 3/8) / (n + 1/4).}}
  }
  Where \eqn{f} is the fraction of values that lies below \eqn{i}, and
  \eqn{n} is the total number values.
}
\examples{

 set.seed(321)
 z <- runif(10, 1, 20)
 eda_fval(z)  # Cleveland's f-values
 eda_fval(z, qtype = 7)  # Algorithm used by the quantile() function.
}
\references{
\itemize{
  \item{\link[stats]{quantile}}
  \item{\link[stats]{ppoints}}}
}
