% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_jitter.R
\name{eda_jitter}
\alias{eda_jitter}
\title{Jitter plot}
\usage{
eda_jitter(
  dat,
  x,
  fac = NULL,
  jitter = 0.05,
  p = 1,
  tukey = FALSE,
  horiz = FALSE,
  stat = mean,
  show.stat = TRUE,
  stat.type = "p",
  stat.col = "firebrick",
  stat.fill = "bisque",
  stat.size = 2,
  stat.pch = 21,
  stat.pad = 1,
  xlab = NULL,
  ylab = NULL,
  grey = 0.6,
  pch = 21,
  p.col = "grey50",
  p.fill = "grey80",
  size = 0.8,
  alpha = 0.8,
  reorder = TRUE,
  show.par = TRUE,
  ylim = NULL
)
}
\arguments{
\item{dat}{Data frame}

\item{x}{Column name assigned to the values}

\item{fac}{Column name assigned to the factor the values are to be
conditioned on}

\item{jitter}{Jittering parameter. A fraction of the group width
(\code{0} to \code{1})}

\item{p}{Power transformation to apply to variable}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation)}

\item{horiz}{plot horizontally (TRUE) or vertically (FALSE)}

\item{stat}{Choice of summary statistic to use when centering the fitted
values around 0. \code{stat} can be either \code{mean} or \code{median}. DO
NOT wrap stat parameter in quotes.}

\item{show.stat}{Boolean determining if a summary statistic of the batches
should be shown.}

\item{stat.type}{Symbol to use to display statistical summary. Can be either
a point, \code{"p"}, or a line, \code{"l"}.}

\item{stat.col}{Symbol color to use to display statistical summary. If
\code{stat.type} is a point, then the color will be passed to its outline
if the point symbol type is 21 through 25.}

\item{stat.fill}{Fill color to use for the point statistical summary if
point symbol type is 21 through 25. Ignored if \code{stat.type = "l"}.}

\item{stat.size}{Size of point statistical summary if \code{stat.type = "p"},
or width of line if \code{stat.type = "l"}.}

\item{stat.pch}{Point statistical summary type (1 through 25). Ignored if
\code{stat.type = "l"}.}

\item{stat.pad}{Fraction to add to the length of the line statistical summary,
Ignore if \code{stat.type = "p"}.}

\item{xlab}{X label for output plot}

\item{ylab}{Y label for output plot}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black)}

\item{pch}{Point symbol type}

\item{p.col}{Color for point symbol}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point size (0-1).}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}

\item{reorder}{Boolean determining if factors have to be reordered based
on \code{stat}.}

\item{show.par}{Boolean determining if power transformation should be
displayed in the plot.}

\item{ylim}{Define custom y axis limits (x limits if \code{horiz=TRUE}.}
}
\value{
{No values are returned}
}
\description{
\code{eda_jitter} creates jitter plot conditioned on one variable.
}
\details{
Function generates jitter plot from a univariate dataset. If no categorical
variable is passed to the function, a single jitter plot is created from the
data. \cr
\cr
By default, the plots are ordered based on the statistic used to generate
the central value.
}
\examples{

# A basic jitter plot for a single variable where the mean is automatically
# fitted to the data
eda_jitter(iris, Sepal.Width)

# A basic jitter plot conditioned on a categorical variable
eda_jitter(iris, Sepal.Width, Species)

# Use line instead of point for summary statistic
eda_jitter(iris, Sepal.Width, Species, stat.type = "l")

# You can apply a transformation to the data. The summary statistic will be
# computed after the data are transformed.
eda_jitter(mtcars, hp, cyl, p = 0)
}
