% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_lm.R
\name{eda_lm}
\alias{eda_lm}
\title{Least Squares regression plot (with optional LOESS line)}
\usage{
eda_lm(
  dat,
  x,
  y,
  x.lab = NULL,
  y.lab = NULL,
  reg = TRUE,
  loe = FALSE,
  lm.col = rgb(1, 0.5, 0.5, 0.8),
  loe.col = rgb(0.73, 0.73, 1, 1),
  stats = FALSE,
  plot.d = list(pch = 20, col = "grey40"),
  ...,
  loess.d = NULL
)
}
\arguments{
\item{dat}{data frame}

\item{x}{column name assigned the x axis}

\item{y}{column name assigned the y axis}

\item{x.lab}{X label for output plot}

\item{y.lab}{Y label for output plot}

\item{reg}{boolean indicating whether a least squares regression line
should be plotted}

\item{loe}{boolean indicating if a loess curve should be fitted}

\item{lm.col}{regression line color}

\item{loe.col}{LOESS curve color}

\item{stats}{boolean indicating if regression summary statistics should be
displayed}

\item{plot.d}{Additional parameters passed to the plot function}

\item{...}{not used}

\item{loess.d}{Additional parameters passed to the loess.smooth function}
}
\description{
\code{eda_lm} generates a scatter plot with a fitted regression line.
 A loess line can  also be added to the plot
 for model comparison. The axes are scaled such that their respective
 standard  deviations match axes unit length.
}
\examples{

eda_lm(mtcars, wt, mpg, plot.d = list(pch=16, col="black"), loe=TRUE,
      loess.d=list(family = "symmetric", span=0.5, degree=2))

eda_lm(mtcars, wt, mpg, plot.d = list(pch=16, col="blue"), loe=TRUE)
}
\seealso{
\code{\link[graphics]{plot}} and \code{\link[stats]{loess.smooth}} functions
}
