% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_lm.R
\name{eda_lm}
\alias{eda_lm}
\title{Least Squares regression plot (with optional LOESS fit)}
\usage{
eda_lm(
  dat,
  x,
  y,
  xlab = NULL,
  ylab = NULL,
  px = 1,
  py = 1,
  tukey = FALSE,
  reg = TRUE,
  w = NULL,
  sd = TRUE,
  grey = 0.6,
  pch = 21,
  p.col = "grey50",
  p.fill = "grey80",
  size = 0.8,
  alpha = 0.8,
  q = FALSE,
  q.val = c(0.16, 0.84),
  q.type = 5,
  loe = FALSE,
  lm.col = rgb(1, 0.5, 0.5, 0.8),
  loe.col = rgb(0.3, 0.3, 1, 1),
  stats = FALSE,
  loess.d = list(family = "symmetric", span = 0.7, degree = 1),
  ...
)
}
\arguments{
\item{dat}{Data frame}

\item{x}{Column assigned to the x axis}

\item{y}{Column assigned to the y axis}

\item{xlab}{X label for output plot}

\item{ylab}{Y label for output plot}

\item{px}{Power transformation to apply to the x-variable}

\item{py}{Power transformation to apply to the y-variable}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation)}

\item{reg}{Boolean indicating whether a least squares regression line should
be plotted}

\item{w}{Weight to pass to regression model}

\item{sd}{Boolean determining if standard deviation lines should be plotted}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black)}

\item{pch}{Point symbol type}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point size (0-1)}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}

\item{q}{Boolean determining if grey quantile boxes should be plotted}

\item{q.val}{F-values to use to define the quantile box parameters. Defaults
to mid 68% of values. If more than 2 f-values are defined, the first two
are used to generate the box.}

\item{q.type}{Quantile type. Defaults to 5 (Cleveland's f-quantile
definition)}

\item{loe}{Boolean indicating if a loess curve should be fitted}

\item{lm.col}{Regression line color}

\item{loe.col}{LOESS curve color}

\item{stats}{Boolean indicating if regression summary statistics should be
displayed}

\item{loess.d}{A list of parameters passed to the \code{loess.smooth}
function. A robust loess is used by default.}

\item{...}{Not used}
}
\value{
Returns a list from the  `lm` output.

\itemize{
  \item \code{a}: Intercept
  \item \code{b}: Slope
  \item \code{residuals}: Regression model residuals}
}
\description{
\code{eda_lm} generates a scatter plot with a fitted regression
  line. A loess line can  also be added to the plot for model comparison. The
  axes are scaled such that their respective standard  deviations match axes
  unit length.
}
\examples{

# Add a regular (OLS) regression model and loess smooth to the data
eda_lm(mtcars, wt, mpg, loe = TRUE)

# Add the inner 68\% quantile to compare the true 68\% of data to the SD
eda_lm(mtcars, wt, mpg, loe = TRUE, q = TRUE)

# Show the IQR box
eda_lm(mtcars, wt, mpg, loe = TRUE, q = TRUE, sd = FALSE, q.val = c(0.25,0.75))

# Fit an OLS to the Income for Female vs Male
df2 <- read.csv("https://mgimond.github.io/ES218/Data/Income_education.csv")
eda_lm(df2, x=B20004013, y = B20004007, xlab = "Female", ylab = "Male",
            loe = TRUE)

# Add the inner 68\% quantile to compare the true 68\% of data to the SD
eda_lm(df2, x = B20004013, y = B20004007, xlab = "Female", ylab = "Male",
            q = TRUE)

# Apply a transformation to x and y axes: x -> 1/3 and y -> log
eda_lm(df2, x = B20004013, y = B20004007, xlab = "Female", ylab = "Male",
            px = 1/3, py = 0, q = TRUE, loe = TRUE)

}
\seealso{
\code{\link[graphics]{plot}} and \code{\link[stats]{loess.smooth}}
  functions
}
