% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_pol.R
\name{eda_pol}
\alias{eda_pol}
\title{Polish two-way tables}
\usage{
eda_pol(
  x,
  row = NULL,
  col = NULL,
  val = NULL,
  stat = median,
  plot = TRUE,
  eps = 0.01,
  maxiter = 5,
  sort = FALSE,
  p = 1,
  tukey = FALSE,
  offset = 1e-05,
  col.quant = FALSE,
  colpal = "RdYlBu",
  adj.mar = FALSE,
  res.size = 1,
  row.size = 1,
  col.size = 1,
  res.txt = TRUE,
  label.txt = TRUE
)
}
\arguments{
\item{x}{A three column data frame}

\item{row}{Name of column assigned to the row effect}

\item{col}{Name of column assigned to the column effect}

\item{val}{Name of column assigned to the response variable}

\item{stat}{Polishing statistic (default is median)}

\item{plot}{Boolean determining if an output plot should be generated}

\item{eps}{Convergence tolerance parameter}

\item{maxiter}{Maximum number of iterations}

\item{sort}{Boolean determining if the effects row/columns should be sorted}

\item{p}{Re-expression power parameter}

\item{tukey}{Boolean determining if Tukey's power transformation should used.
If FALSE, the Box-Cox transformation is adopted.}

\item{offset}{Offset to add to values if at leat one value is 0 and the power
is negative}

\item{col.quant}{Boolean determining if a quantile classification scheme
should be used}

\item{colpal}{Color palette to adopt}

\item{adj.mar}{Boolean determining if margin width needs to accomodate labels}

\item{res.size}{Size of residual values in plot [0-1]}

\item{row.size}{Size of row effect values in plot [0-1]}

\item{col.size}{Size of column effect values in plot [0-1]}

\item{res.txt}{Boolean determining if values should be added to plot}

\item{label.txt}{Boolean determining if margin and column labels should be
plotted}
}
\value{
A list of class \code{eda_polish} with the following named
  components:

\itemize{
\item \code{long} The median polish residuals with three columns: Column levels,
  row levels and residual values.
\item \code{wide} The median polish residuals table in wide form.
\item \code{row}  Row effects table
\item \code{col}  Column effects table
\item \code{global}  Overall value (common value)
\item \code{iter}  Number of iterations before polish stabilizes.
\item \code{cv}  Table of residuals, row effects, column effects and CV values in long form.
\item \code{power}  Transformation power applied to values prior to polishing.
\item \code{IQ_row}  Ratio between interquartile row effect values and 80th quantile of residuals.
\item \code{IQ_col}  Ratio between interquartile column effect values and 80th quantile of residuals.}
}
\description{
\code{eda_pol} Polishes two-way tables using median, means, or
  any customizable functions.
}
\details{
The function performs a polish on a two way table. By default, it
  applies a median polish, but other statistical summaries such as the mean
  can be passed to the function via the \code{stat = } parameter. The
  function returns a list of row/column effects along with global and
  residual values. It will also generate a colored table if \code{plot =
  TRUE}.
}
\examples{
df <- data.frame(region =  rep( c("NE", "NC", "S", "W"), each = 5),
edu = rep( c("ed8", "ed9to11", "ed12", "ed13to15", "ed16"), 4),
perc = c(25.3, 25.3, 18.2, 18.3, 16.3, 32.1, 29, 18.8,
        24.3, 19, 38.8, 31, 19.3, 15.7, 16.8, 25.4, 21.1, 20.3, 24, 17.5))

eda_pol(df, row = "region", col = "edu", val = "perc", plot = FALSE)
}
