% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_qqmat.R
\name{eda_qqmat}
\alias{eda_qqmat}
\title{Matrix QQ plot}
\usage{
eda_qqmat(
  dat,
  x,
  fac,
  p = 1L,
  tukey = FALSE,
  q.type = 5,
  diag = TRUE,
  xylim = NULL,
  resid = FALSE,
  stat = mean,
  plot = TRUE,
  grey = 0.6,
  pch = 21,
  p.col = "grey50",
  p.fill = "grey80",
  size = 0.5,
  tic.size = 0.7,
  alpha = 0.8,
  q = TRUE,
  show.med = TRUE,
  inner = 0.75,
  ...
)
}
\arguments{
\item{dat}{Data frame}

\item{x}{Column name assigned to the values}

\item{fac}{Column name assigned to the factor the values are to be
conditioned on}

\item{p}{Power transformation to apply to both sets of values.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation).}

\item{q.type}{An integer between 1 and 9 selecting one of the nine quantile
algorithms. (See \code{quantile}tile function).}

\item{diag}{Boolean determining if upper diagonal should be plotted}

\item{xylim}{X and Y axes limits.}

\item{resid}{Boolean determining if residuals should be plotted. Residuals
are computed using the \code{stat} parameter.}

\item{stat}{Statistic to use if residuals are to be computed. Currently
\code{mean} (default) or \code{median}.}

\item{plot}{Boolean determining if plot should be generated.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{pch}{Point symbol type.}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point symbol size (0-1)}

\item{tic.size}{Size of tic symbols (defaults to 0.8)}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}

\item{q}{Boolean determining if grey quantile boxes should be plotted.}

\item{show.med}{Boolean determining if median lines should be drawn.}

\item{inner}{Fraction of mid-values to display in the quantile box. Defaults
to the inner 75\% of values.}

\item{...}{Not used}
}
\value{
Returns a list with the following components:

\itemize{
  \item \code{data}: List with input \code{x} and \code{y} values for each
  group.
  May be interpolated to smallest quantile batch.
  Values will reflect power transformation defined in \code{p}.
  \item \code{p}: Re-expression applied to original values.}
}
\description{
\code{eda_qqmat} Generates a matrix of empirical QQ plots
}
\details{
When the function is used to generate an empirical QQ plots, the plot
  will displays the IQR via grey boxes for both x and y values. The box
  widths can be changed via the  \code{inner} argument.  The middle dashed
  line represents each batch's median value. \cr
  \cr
}
\examples{

singer <- lattice::singer
eda_qqmat(singer, height, voice.part)
eda_qqmat(singer, height, voice.part, diag = FALSE)
eda_qqmat(singer, height, voice.part, diag = FALSE, resid = TRUE, q = FALSE)
eda_qqmat(mtcars, mpg, cyl,resid = TRUE, q = FALSE)
eda_qqmat(iris, Petal.Length, Species, resid = TRUE, p = 0)

}
\references{
\itemize{
  \item John M. Chambers, William S. Cleveland, Beat Kleiner, Paul A. Tukey.
  Graphical Methods for Data Analysis (1983)}
}
