% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_qqmat.R
\name{eda_qqmat}
\alias{eda_qqmat}
\title{QQ plot matrix}
\usage{
eda_qqmat(
  dat,
  x,
  fac,
  p = 1L,
  tukey = FALSE,
  q.type = 5,
  diag = TRUE,
  xylim = NULL,
  resid = FALSE,
  stat = mean,
  plot = TRUE,
  grey = 0.6,
  pch = 21,
  p.col = "grey40",
  p.fill = "grey60",
  size = 1,
  text.size = 1,
  tail.pch = 21,
  tail.p.col = "grey70",
  tail.p.fill = NULL,
  tic.size = 0.7,
  alpha = 0.8,
  q = FALSE,
  tails = TRUE,
  med = TRUE,
  inner = 0.75,
  ...
)
}
\arguments{
\item{dat}{Data frame.}

\item{x}{Continuous variable.}

\item{fac}{Categorical variable.}

\item{p}{Power transformation to apply to the continuous variable.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation).}

\item{q.type}{An integer between 1 and 9 selecting one of the nine quantile
algorithms. (See \code{quantile}tile function).}

\item{diag}{Boolean determining if both upper and lower triangular matrix
should be plotted. If set to \code{FALSE}, only the lower triangular matrix
is plotted.}

\item{xylim}{X and Y axes limits.}

\item{resid}{Boolean determining if residuals should be plotted. Residuals
are computed using the \code{stat} parameter.}

\item{stat}{Statistic to use if residuals are to be computed. Currently
\code{mean} (default) or \code{median}.}

\item{plot}{Boolean determining if plot should be generated.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{pch}{Point symbol type.}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point symbol size (0-1).}

\item{text.size}{Size for category text in diagonal box.}

\item{tail.pch}{Tail-end point symbol type (See \code{tails}).}

\item{tail.p.col}{Tail-end color for point symbol (See \code{tails}).}

\item{tail.p.fill}{Tail-end point fill color passed to \code{bg}
(Only used for \code{tail.pch} ranging from 21-25).}

\item{tic.size}{Size of tic labels (defaults to 0.8).}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}

\item{q}{Boolean determining if grey box highlighting the \code{inner}
region should be displayed.}

\item{tails}{Boolean determining if points outside of the \code{inner} region
 should be symbolized differently. Tail-end points are symbolized via the
\code{tail.pch}, \code{tail.p.col} and \code{tail.p.fill} arguments.}

\item{med}{Boolean determining if median lines should be drawn.}

\item{inner}{Fraction of mid-values to highlight in \code{q} or \code{tails}.
Defaults to the inner 75\% of values.}

\item{...}{Not used}
}
\value{
Returns a list with the following components:

\itemize{
  \item \code{data}: List with input \code{x} and \code{y} values for each
  group. May be interpolated to smallest quantile batch if batch sizes
  don't match. Values will reflect power transformation defined in \code{p}.
  \item \code{p}: Transformation applied to original values.}
}
\description{
\code{eda_qqmat} Generates a matrix of empirical QQ plots
}
\details{
The function will generate an empirical QQ plot matrix from a
dataframe of continuous values and matching categories. The function is
designed to place emphasis on the mid portion of the data. The mid portion
range is defined by \code{inner} (the inner fraction of the data). By default,
the points outside of the mid portion of the data are symbolized differently.
You can also highlight the mid region in light grey by setting
\code{q = TRUE}. The median of both batches are shown in vertical and
horizontal dashed lines. For a plain vanilla QQ plot matrix you can remove
all guides by setting \code{tails = FALSE} and  \code{mid = FALSE}.

The QQ plot matrix is most effective in comparing residuals after the data
are fitted by the mean or median. To plot the residuals, set
\code{resid=TRUE}. By default, the \code{mean} is used. You can change the
statistic to the median by setting \code{stat=median}.

The function also allows for batch transformation of values via the
\code{p} argument. The transformation is applied to the data prior to
computing the residuals.
}
\examples{

# Default output
singer <- lattice::singer
eda_qqmat(singer, height, voice.part)

# Limit to lower triangular matrix
eda_qqmat(singer, height, voice.part, diag = FALSE)

# Plot residuals after fitting mean model
eda_qqmat(singer, height, voice.part, resid = TRUE)

# Generate plain vanilla QQ plot matrix
eda_qqmat(mtcars, mpg, cyl,resid = TRUE, tails = FALSE, med = FALSE)

# Log transform the data, then plot the residuals after fitting the mean model
eda_qqmat(iris, Petal.Length, Species, resid = TRUE, p = 0)

# Fit the median model instead of the mean
eda_qqmat(iris, Petal.Length, Species, resid = TRUE, p = 0, stat = median)

# Fill inner region with grey boxes
eda_qqmat(iris, Petal.Length, Species, resid = TRUE, q = TRUE, p = 0)

# Change tail point symbol
eda_qqmat(iris, Petal.Length, Species, resid = TRUE, p = 0, tail.pch = 3)

# Change inner region point symbols to dark orange and reduce size of all
# point symbols
eda_qqmat(iris, Petal.Length, Species, resid = TRUE, p = 0, size = 0.8,
          tail.pch = 3, p.fill = "darkorange")
}
\references{
\itemize{
  \item John M. Chambers, William S. Cleveland, Beat Kleiner, Paul A. Tukey.
  Graphical Methods for Data Analysis (1983)}
}
