% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_rfs.R
\name{eda_rfs}
\alias{eda_rfs}
\title{Residual-Fit Spread Plot}
\usage{
eda_rfs(
  dat,
  x = NULL,
  grp = NULL,
  p = 1L,
  tukey = FALSE,
  show.par = TRUE,
  stat = mean,
  grey = 0.7,
  pch = 21,
  p.col = "grey50",
  p.fill = "grey80",
  inner = 0.6826,
  q = FALSE,
  size = 0.8,
  alpha = 0.7,
  ylim = NULL
)
}
\arguments{
\item{dat}{An eda_lm model, an lm model or a dataframe of univariate data.}

\item{x}{Column of values if \code{dat} is a dataframe, ignored otherwise.}

\item{grp}{Column of grouping variables if \code{dat} is a dataframe, ignored
otherwise.}

\item{p}{Power transformation to apply to univariate data. Ignored if linear
model is passed to function.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(\code{TRUE}) or if a Box-Cox transformation should be adopted (\code{FALSE}).}

\item{show.par}{Boolean determining if the power transformation used with the
data should be displayed in the plot's upper-right corner.}

\item{stat}{Choice of summary statistic to use when centering the fitted
values around 0. The \code{stat} parameter is also used for fitting
univariate values (i.e. for summarizing groups). \code{stat} can be either
\code{mean} or \code{median}.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{pch}{Point symbol type.}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{inner}{Fraction of values that should be captured by the inner color
band of the normal and density plots. Defaults to 0.6826 (inner 68\% of
values).}

\item{q}{Boolean determining if grey quantile boxes should be plotted.}

\item{size}{Point size (0-1)}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}

\item{ylim}{Define custom y axis limits.}
}
\description{
\code{eda_rfs} generates a Cleveland residual-fit spread plot
   for univariate or bivariate data.
}
\examples{

# Generate a basic residual-fit spread plot
eda_rfs(mtcars,mpg, cyl)

# Add inner 68.4\% region to residuals (grey boxes in plot)
# Vertical grey box shows matching y-values
eda_rfs(mtcars,mpg, cyl, q = TRUE)

# Change guide to encompass mid 75\% of residual values
eda_rfs(mtcars,mpg, cyl, q = TRUE, inner = 0.75)

# Use median instead of the mean to compute group summaries and to
# recenter the fitted values around 0.
eda_rfs(mtcars,mpg, cyl, stat = median)

# Apply power transformation of -1 to mpg. Defaults to box-cox method.
eda_rfs(mtcars,mpg, cyl, p = -1)

# Generate rfs plot for bivariate model output. Model can be generated from
# lm(), eda_lm() or eda_rline()
M1 <- lm(hp ~ mpg, mtcars)
eda_rfs(M1,q =TRUE)

M2 <- eda_lm(mtcars, mpg, hp)
eda_rfs(M2,q =TRUE)

M3 <- eda_rline(mtcars, mpg, hp)
eda_rfs(M3, q =TRUE)
}
\references{
\itemize{
  \item William S. Cleveland. Visualizing Data. Hobart Press (1993)}
}
