% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_rline.R
\name{eda_rline}
\alias{eda_rline}
\title{Tukey's resistant line}
\usage{
eda_rline(dat, x, y)
}
\arguments{
\item{dat}{data frame}

\item{x}{column name assigned the x axis}

\item{y}{column name assigned the y axis}
}
\value{
Outputs a plot showing the three point summary as well as a list of
parameters:


\itemize{
  \item \code{a}: Intercept
  \item \code{b}: Slope
  \item \code{res}: Residuals sorted on x-values order
  \item \code{x}: Sorted x values
  \item \code{y}: y values following sorted x-values order
  \item \code{xmed}: Median x values for each third
  \item \code{ymed}: Median y values for each third
  \item \code{index}: Index of sorted x values defining upper boundaries of
                     each thirds}
}
\description{
\code{eda_rline} is an R implementation of Hoaglin, Mosteller and Tukey's
 resistant line technique outlined in chapter 5 of "Understanding Robust
 and Exploratory Data Analysis" (Wiley, 1983).
}
\details{
Bits and pieces of the RLIN.F FORTRAN code in Velleman et. al's book were
used in helping implement some of the subroutines.\cr
Note that this function has only been tested with a subset of datasets. It is far from
being fully vetted. So use with caution!

\emph{References:
  \itemize{
   \item Applications, Basics and Computing of Exploratory Data Analysis,
   by P.F. Velleman and D.C. Hoaglin
   (available at http://dspace.library.cornell.edu/handle/1813/78)
   \item Understanding robust and exploratory data analysis, by D.C. Hoaglin,
   F. Mosteller and J.W. Tukey}}
}
\examples{

# This first test is with breast cancer data from "ABC's of EDA" page 127.
# The final model should look like:  Y = -46.19 + 2.89X

r.lm <- eda_rline(neoplasms, Temp, Mortality)
r.lm

# Check output
OP <- par( mfrow = c(2,1))
  plot(Mortality ~ Temp, neoplasms)
  mtext(sprintf("y = \%f + (\%f)x", r.lm$a, r.lm$b ))
  abline(a = r.lm$a, b = r.lm$b, col="red")
  abline( lm(Mortality ~ Temp, neoplasms), col="grey", lty=3)
  points(cbind(r.lm$xmed,r.lm$ymed), pch =16, col="red")
  abline(v= r.lm$x[r.lm$index],lty=3)
  plot(r.lm$res ~ r.lm$x)
  abline( h = 0, lty=3)
par(OP)

# This next example compares children height to age
r.lm    <- eda_rline(age_height, Months, Height)

OP <- par( mfrow = c(2,1))
 plot(Height ~ Months, age_height, xlab="Age (months)", ylab="Height (cm)")
 mtext(sprintf("y = \%f + (\%f)x", r.lm$a, r.lm$b ))
 abline(a = r.lm$a, b = r.lm$b, col="red")
 abline( lm(Height ~ Months, age_height), col="grey", lty=3)
 points(cbind(r.lm$xmed,r.lm$ymed), pch =16, col="red")
 abline(v= r.lm$x[r.lm$index],lty=3)
 plot(r.lm$res ~ r.lm$x)
 abline( h = 0, lty=3)
par(OP)

# Andrew Siegel's pathological 9-point data set

r.lm <- eda_rline(nine_point, X, Y)

OP <- par( mfrow = c(2,1))
plot(Y ~ X, nine_point, xlab="Age (months)", ylab="Height (cm)")
   mtext(sprintf("y = \%f + (\%f)x", r.lm$a, r.lm$b ))
   abline(a = r.lm$a, b = r.lm$b, col="red")
   abline( lm(Y ~ X, nine_point), col="grey", lty=3)
   points(cbind(r.lm$xmed,r.lm$ymed), pch =16, col="red")
   abline(v= r.lm$x[r.lm$index],lty=3)
   plot(r.lm$res ~ r.lm$x)
   abline( h = 0, lty=3)
par(OP)

}
