% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_rline.R
\name{eda_rline}
\alias{eda_rline}
\title{Tukey's resistant line}
\usage{
eda_rline(dat, x, y)
}
\arguments{
\item{dat}{Data frame}

\item{x}{Column assigned to the x axis}

\item{y}{Column assigned to the y axis}
}
\value{
Returns a list of class \code{eda_rline}with the following named
  components:

\itemize{
  \item \code{a}: Intercept
  \item \code{b}: Slope
  \item \code{res}: Residuals sorted on x-values
  \item \code{x}: Sorted x values
  \item \code{y}: y values following sorted x-values
  \item \code{xmed}: Median x values for each third
  \item \code{ymed}: Median y values for each third
  \item \code{index}: Index of sorted x values defining upper boundaries of
                     each thirds
  \item \code{xlab}: X label name
  \item \code{ylab}: Y label name
  \item \code{iter}: Number of iterations}
}
\description{
\code{eda_rline} is an R implementation of Hoaglin, Mosteller
  and Tukey's resistant line technique outlined in chapter 5 of
  "Understanding Robust and Exploratory Data Analysis" (Wiley, 1983).
}
\details{
This is an R implementation of the \code{RLIN.F} FORTRAN code in
  Velleman et. al's book. This function fits a robust line using a
  three-point summary strategy whereby the data are split into three equal
  length groups along the x-axis and a line is fitted to the medians defining
  each group via an iterative process. This function should mirror the
  built-in \code{stat::line} function in its fitting strategy but it outputs
  additional parameters.

  \cr See the accompanying vignette \code{Resistant Line} for a detailed
  breakdown of the resistant line technique.
}
\examples{

# This first example fits a resistant line to the neoplasms data

M <- eda_rline(neoplasms, Temp, Mortality)
plot(Mortality~Temp, neoplasms, pch=20)
abline(a=M$a, b=M$b, col=rgb(1,0,0,0.5))

# This second example uses breast cancer data from "ABC's of EDA" page 127.
# The output model's  parameters should closely match:  Y = -46.19 + 2.89X
# The plots shows the original data with a fitted resistant line (red)
# and a regular lm fitted line (dashed grey), and the modeled residuals.
# The 3-point summary dots are shown in red.

r.lm <- eda_rline(neoplasms, Temp, Mortality)
r.lm

# Check output
OP <- par( mfrow = c(2,1))
  plot(Mortality ~ Temp, neoplasms)
  mtext(sprintf("y = \%f + (\%f)x", r.lm$a, r.lm$b ))
  abline(a = r.lm$a, b = r.lm$b, col="red")
  abline( lm(Mortality ~ Temp, neoplasms), col="grey", lty=3)
  points(cbind(r.lm$xmed,r.lm$ymed), pch =16, col="red")
  abline(v= r.lm$x[r.lm$index],lty=3)
  plot(r.lm$res ~ r.lm$x)
  abline( h = 0, lty=3)
par(OP)

# This next example compares children height to age.
# The plots shows the original data with a fitted resistant line (red)
# and a regular lm fitted line (dashed grey), and the modeled residuals.
# The 3-point summary dots are shown in red.
r.lm    <- eda_rline(age_height, Months, Height)

OP <- par( mfrow = c(2,1))
 plot(Height ~ Months, age_height, xlab="Age (months)", ylab="Height (cm)")
 mtext(sprintf("y = \%f + (\%f)x", r.lm$a, r.lm$b ))
 abline(a = r.lm$a, b = r.lm$b, col="red")
 abline( lm(Height ~ Months, age_height), col="grey", lty=3)
 points(cbind(r.lm$xmed,r.lm$ymed), pch =16, col="red")
 abline(v= r.lm$x[r.lm$index],lty=3)
 plot(r.lm$res ~ r.lm$x)
 abline( h = 0, lty=3)
par(OP)

# Andrew Siegel's pathological 9-point data set
r.lm <- eda_rline(nine_point, X, Y)

OP <- par( mfrow = c(2,1))
plot(Y ~ X, nine_point, xlab="Age (months)", ylab="Height (cm)")
   mtext(sprintf("y = \%f + (\%f)x", r.lm$a, r.lm$b ))
   abline(a = r.lm$a, b = r.lm$b, col="red")
   abline( lm(Y ~ X, nine_point), col="grey", lty=3)
   points(cbind(r.lm$xmed,r.lm$ymed), pch =16, col="red")
   abline(v= r.lm$x[r.lm$index],lty=3)
   plot(r.lm$res ~ r.lm$x)
   abline( h = 0, lty=3)
par(OP)

}
\references{
\itemize{
   \item Velleman, P. F., and D. C. Hoaglin. 1981. Applications, Basics and Computing of Exploratory Data Analysis. Boston: Duxbury Press.
   \item D. C. Hoaglin, F. Mosteller, and J. W. Tukey. 1983. Understanding Robust and Exploratory Data Analysis. Wiley.}
}
