% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_rline.R
\name{eda_rline}
\alias{eda_rline}
\title{Tukey's resistant line}
\usage{
eda_rline(dat, x, y, px = 1, py = 1, tukey = FALSE, maxiter = 20)
}
\arguments{
\item{dat}{Data frame.}

\item{x}{Column assigned to the x axis.}

\item{y}{Column assigned to the y axis.}

\item{px}{Power transformation to apply to the x-variable.}

\item{py}{Power transformation to apply to the y-variable.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted.
(FALSE adopts a Box-Cox transformation)}

\item{maxiter}{Maximum number of iterations to run.}
}
\value{
Returns a list of class \code{eda_rline} with the following named
components:

\itemize{
\item \code{a}: Intercept
\item \code{b}: Slope
\item \code{residuals}: Residuals sorted on x-values
\item \code{x}: Sorted x values
\item \code{y}: y values following sorted x-values
\item \code{xmed}: Median x values for each third
\item \code{ymed}: Median y values for each third
\item \code{index}: Index of sorted x values defining upper boundaries of
each thirds
\item \code{xlab}: X label name
\item \code{ylab}: Y label name
\item \code{iter}: Number of iterations
\item \code{fitted.values}: Fitted values}
}
\description{
\code{eda_rline} is an R implementation of Hoaglin, Mosteller
and Tukey's resistant line technique outlined in chapter 5 of
"Understanding Robust and Exploratory Data Analysis" (Wiley, 1983).
}
\details{
This is an R implementation of the \code{RLIN.F} FORTRAN code in
Velleman et. al's book. This function fits a robust line using a
three-point summary strategy whereby the data are split into three equal
length groups along the x-axis and a line is fitted to the medians defining
each group via an iterative process. This function should mirror the
built-in \code{stat::line} function in its fitting strategy but it outputs
additional parameters.

\cr See the accompanying \href{../articles/RLine.html}{resistant line
article} for a detailed breakdown of the resistant line technique.
}
\examples{

# This first example uses breast cancer data from "ABC's of EDA" page 127.
# The output model's  parameters should closely match:  Y = -46.19 + 2.89X
# The plots shows the original data with a fitted resistant line (red)
# and a regular lm fitted line (dashed line), and the modeled residuals.
# The 3-point summary dots are shown in red.

M <- eda_rline(neoplasms, Temp, Mortality)
M

# Plot the output (red line is the resistant line)
plot(M)

# Add a traditional OLS regression line (dashed line)
abline(lm(Mortality ~ Temp, neoplasms), lty = 3)

# Plot the residuals
plot(M, type = "residuals")

# This next example uses Andrew Siegel's pathological 9-point dataset to test
# for model stability when convergence cannot be reached.
M <- eda_rline(nine_point, X, Y)
plot(M)

}
\references{
\itemize{
\item Velleman, P. F., and D. C. Hoaglin. 1981. Applications, Basics and Computing of Exploratory Data Analysis. Boston: Duxbury Press.
\item D. C. Hoaglin, F. Mosteller, and J. W. Tukey. 1983. Understanding Robust and Exploratory Data Analysis. Wiley.}
}
