% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_sl.R
\name{eda_sl}
\alias{eda_sl}
\title{Spread-location and spread-level plots}
\usage{
eda_sl(
  dat,
  x = NULL,
  fac = NULL,
  type = "location",
  p = 1,
  tukey = FALSE,
  base = exp(1),
  sprd = "frth",
  jitter = 0.01,
  robust = TRUE,
  loess.d = list(family = "symmetric", degree = 1, span = 1),
  label = TRUE,
  label.col = "lightsalmon",
  xlab = NULL,
  ylab = NULL,
  labelxbuff = 0.05,
  labelybuff = 0.05,
  show.par = FALSE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{dat}{Dataframe of univariate data or a linear model.}

\item{x}{Continuous variable column (ignored if \code{dat} is a linear model).}

\item{fac}{Categorical variable column (ignored if \code{dat} is a linear
model).}

\item{type}{s-l plot type. \code{"location"} = spread-location,
\code{"level"} = spread-level (only for univariate data).
\code{"dependence"} = spread-dependence (only for bivariate model input).}

\item{p}{Power transformation to apply to variable. Ignored if input is a
linear model.}

\item{tukey}{Logical; Determines if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation).}

\item{base}{Base used with the \code{log()} function if \code{px} or \code{py}
is \code{0}.}

\item{sprd}{Choice of spreads used in the spread-versus-level plot (i.e.
when \code{type = "level"}). Either
interquartile, \code{sprd = "IQR"} or
fourth-spread, \code{sprd = "frth"} (default).}

\item{jitter}{Jittering parameter for the spread-location plot. A fraction of
the range of location values.}

\item{robust}{Logical; Indicates if robust regression should be used on the
spread-level plot.}

\item{loess.d}{Arguments passed to the internal loess function. Applies only
to the bivariate model s-l plots and the spread-level plot.}

\item{label}{Logical; Determines if group labels are to be added to the
spread-location plot.}

\item{label.col}{Color assigned to group labels (only applicable if
\code{type = location}).}

\item{xlab}{X label for output plot.}

\item{ylab}{Y label for output plot.}

\item{labelxbuff}{Buffer to add to the edges of the plot to make room for
the labels in a spread-location plot. Value is a fraction of the plot width.}

\item{labelybuff}{Buffer to add to the top of the plot to make room for
the labels in a spread-location plot. Value is a fraction of the plot width.}

\item{show.par}{Boolean determining if the power transformation applied to
the data should be displayed.}

\item{plot}{Logical; Determines if plot should be generated.}

\item{...}{
  Arguments passed on to \code{\link[=.eda_plot_xy]{.eda_plot_xy}}
  \describe{
    \item{\code{y}}{A numeric vector or column name in \code{dat} for the y-axis.}
    \item{\code{px}}{Power transformation used in the input data to display if
\code{show.par = TRUE}.}
    \item{\code{py}}{Power transformation used in the input data to display if
\code{show.par = TRUE}.}
    \item{\code{raw_tick}}{Logical. If \code{TRUE}, original (untransformed) equally
spaced tick values are displayed on the re-expressed axes.}
    \item{\code{xlim}}{X-axis range.}
    \item{\code{ylim}}{Y-axis range.}
    \item{\code{reg}}{Logical; whether to fit and display a regression line.}
    \item{\code{poly}}{Integer; regression model polynomial degree (defaults to 1 for
linear model).}
    \item{\code{rlm.d}}{List; parameters for \code{MASS::rlm}, (e.g., \code{list(psi =
"psi.bisquare")}).}
    \item{\code{w}}{Optional numeric vector of weights for regression.}
    \item{\code{lm.col}}{Regression line color.}
    \item{\code{lm.lw}}{Numeric; Regression line width.}
    \item{\code{lm.lty}}{Numeric; Regression line type.}
    \item{\code{sd}}{Logical; whether to show ±1 SD lines.}
    \item{\code{mean.l}}{Logical; whether to show x and y mean reference lines.}
    \item{\code{asp}}{Logical; whether to preserve the aspect ratio (ignored if
\code{square = FALSE}).}
    \item{\code{square}}{Logical; whether to create a square plotting window.}
    \item{\code{grey}}{Numeric between \code{0-1}; controls grayscale background elements
(\code{0 = black}, \code{1 = white}).}
    \item{\code{pch}}{Integer; point symbol.}
    \item{\code{p.col}}{Point border color.}
    \item{\code{p.fill}}{Point fill color.}
    \item{\code{size}}{Point size.}
    \item{\code{alpha}}{Point transparency level (0 = 100\\\% transparent, 1 = 100\\\% opaque).}
    \item{\code{q}}{Logical; whether to draw inner quantile boxes (quantile shading).}
    \item{\code{q.type}}{Integer; type of quantile calculation (see \code{quantile}).}
    \item{\code{inner}}{Numeric; defines the inner fraction of values to highlight
with quantile shading.}
    \item{\code{qcol}}{Fill color of quantile shading.}
    \item{\code{loe}}{Logical; whether to plot loess smooth line.}
    \item{\code{loe.lw}}{Numeric; Loess smooth line width.}
    \item{\code{loe.col}}{Loess smooth color.}
    \item{\code{loe.lty}}{Numeric; Loess smooth line type.}
    \item{\code{stats}}{Logical; if \code{TRUE}, displays model statistics (R², β, p-value).}
    \item{\code{stat.size}}{Text size for \code{stats} plot display.}
    \item{\code{hline}}{Numeric; location(s) of additional horizontal reference lines.
Can be passed via  the \code{c()} function.}
    \item{\code{vline}}{Numeric; location(s) of additional vertical reference lines.
Can be passed via the \code{c()} function.}
  }}
}
\value{
Returns a dataframe of level and spread values.
}
\description{
The \code{eda_sl} function generates William Cleveland's
\strong{spread-location plot} for univariate and bivariate data. The function will also
generate Tukeys' \strong{spread-level plot}.
}
\details{
The function generates a few variations of the spread-location/spread-level
plots depending on the data input type and parameter passed to the
\code{type} argument. The residual spreads are mapped to the y-axis and the
levels are mapped to the x-axis. Their values are computed as follows:
\itemize{

\item \code{type = "location"} (univariate data):\cr\cr
William Cleveland's spread-location plot applied to univariate
data.\cr
\eqn{\ spread = \sqrt{|residuals|}} \cr
\eqn{\ location = medians}

\item \code{type = "level"} (univariate data):\cr\cr
Tukey's spread-level plot (aka spread-versus-level plot, Hoaglin et al.,
p 260). If the pattern is close to linear, the plot can help find a power
transformation that will help stabilize the spread in the data by
subtracting one from the fitted slope. This option outputs the slope of
the fitted line in the console. A loess is added to assess linearity.
By default, the fourth spread is used to define the spread. Alternatively,
the IQR can be used by setting \code{spread = "IQR"}. The output will be
nearly identical except for small datasets where the two methods may
diverge slightly in output.\cr
\eqn{\ spread = log(fourth\ spread(residuals))} \cr
\eqn{\ location = log(medians)}

\item \code{type = "location"} if input is a model of class \code{lm},
\code{eda_lm} or \code{eda_rline}:\cr\cr
William Cleveland's spread-location plot (aka scale-location plot)
applied to residuals of a linear model.\cr
\eqn{\ spread = \sqrt{|residuals|}} \cr
\eqn{\ location = fitted\ values}

\item \code{type = "dependence"} if input is a model of class \code{lm},
\code{eda_lm} or \code{eda_rline}:\cr\cr
William Cleveland's spread-location plot applied to residuals of
a linear model.\cr
\eqn{\ spread = \sqrt{|residuals|}} \cr
\eqn{\ dependence = x\ variable}
}
}
\examples{
cars <- MASS::Cars93
# Cleveland's spread-location plot applied to univariate data
eda_sl(cars, MPG.city, Type)

# You can specify the exact form of the spread on the y-axis
# via the ylab argument
eda_sl(cars, MPG.city, Type, ylab = expression(sqrt(abs(residuals))) )

# The function can also generate Tukey's spread-level plot to identify a
# power transformation that can stabilize spread across fitted values
# following power = 1 - slope
eda_sl(cars, MPG.city, Type, type = "level")

# A slope of around 3 is computed from the s-l plot, therefore, a suggested
# power is 1 - 3 = -2. We can apply a power transformation within the
# function via the p argument. By default, a Box-Cox transformation method
# is adopted.
eda_sl(cars, MPG.city, Type, p = -2)

# Spread-location plot can also be generated from residuals of a linear model
M1 <- lm(mpg ~ hp, mtcars)
eda_sl(M1)

# Spread can be compared to X instead of fitted value
eda_sl(M1, type = "dependence")
}
\references{
\itemize{
\item Understanding Robust and Exploratory Data Analysis, Hoaglin,
David C., Frederick Mosteller, and John W. Tukey, 1983.
\item William S. Cleveland. Visualizing Data. Hobart Press (1993)
}
}
