% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_sl.R
\name{eda_sl}
\alias{eda_sl}
\title{Tukey's spread-level function}
\usage{
eda_sl(dat, x, y)
}
\arguments{
\item{x}{Categorical variable}

\item{y}{Continuous variable}

\item{all}{generate upper, lower and mid summaries if TRUE or}
}
\description{
\code{eda_sl} The spread-level function generates a spread-level
   table from a univariate dataset
}
\details{
\itemize{
  \item This function depends on the dplyr package.
  \item Note that this function is not to be confused with Bill Cleveland's
  spread-location function.\cr
  \item If x is not categorical, the output will produce many or all NA's.
  }
}
\examples{
sl <- eda_sl(iris, Species, Sepal.Length)
plot(sprd ~ med, sl, pch=16)
}
\references{
Exploratory Data Analysis, John Tukey, 1973.
}
