% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_sl.R
\name{eda_sl}
\alias{eda_sl}
\title{Tukey's spread-level function}
\usage{
eda_sl(
  dat,
  x,
  fac,
  p = 1,
  tukey = FALSE,
  sprd = "frth",
  plot = TRUE,
  grey = 0.6,
  pch = 21,
  p.col = "grey50",
  p.fill = "grey80",
  size = 1,
  alpha = 0.8
)
}
\arguments{
\item{dat}{Dataframe}

\item{x}{Continuous variable column}

\item{fac}{Categorical variable column}

\item{p}{Power transformation to apply to variable}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(FALSE adopts a Box-Cox transformation)}

\item{sprd}{Choice of spreads. Either interquartile, \code{sprd = "IQR"} or
fourth-spread, \code{sprd = "frth"} (default).}

\item{plot}{Boolean determining if plot should be generated.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{pch}{Point symbol type.}

\item{p.col}{Color for point symbol.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point size (0-1)}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only
applicable if \code{rgb()} is not used to define point colors.}
}
\value{
Returns a dataframe of level and spreads.
}
\description{
The \code{eda_sl} function generates a spread-level table from a
univariate dataset.
}
\details{
\itemize{
  \item Note that this function is not to be confused with William Cleveland's
  spread-location function. A description of this plot can be found on page
  77 of *Hoaglan et. al's* book.\cr
  \item If \code{fac} is not categorical, the output will produce many or all
   \code{NA}'s.
  \item On page 59, Hoaglan et. al define the fourth-spread as the the range
  defined by the upper fourth and lower fourth. The \code{eda_lsum} function is used
  to compute the upper/lower fourths.
  }
}
\examples{
sl <- eda_sl(iris, Petal.Length, Species)

# The output can be passed to a model fitting function like eda_lm
# The output slope can be used to help identify a power transformation
# The suggested power transformation is 1 - slope.
eda_lm(sl, Level, Spread)

}
\references{
Understanding Robust and Exploratory Data Analysis, Hoaglin,
David C., Frederick Mosteller, and John W. Tukey, 1983.
}
