% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_sl.R
\name{eda_sl}
\alias{eda_sl}
\title{Tukey's spread-level function}
\usage{
eda_sl(dat, x, y, sprd = "frth")
}
\arguments{
\item{dat}{Dataframe}

\item{x}{Categorical variable column}

\item{y}{Continuous variable column}

\item{sprd}{Choice of spreads. Either interquartile, \code{sprd = "IQR"} or
fourth-spread, \code{sprd = "frth"} (default).}
}
\value{
Returns a dataframe of level and spreads.
}
\description{
The \code{eda_sl} function generates a spread-level table from a
univariate dataset.
}
\details{
\itemize{
  \item Note that this function is not to be confused with Bill Cleveland's
  spread-location function.\cr
  \item If x is not categorical, the output will produce many or all NA's.
  \item On page 59, Hoaglan et. al define the fourth-spread as the the range
  defined by the upper fourth and lower fourth. The \code{eda_lsum} function is used
  to compute the upper/lower fourths.
  }
}
\examples{
sl <- eda_sl(iris, Species, Sepal.Length)
plot(spread ~ level, sl, pch=16)
}
\references{
Understanding Robust and Exploratory Data Analysis, Hoaglin,
David C., Frederick Mosteller, and John W. Tukey, 1983.
}
