% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_unipow.R
\name{eda_unipow}
\alias{eda_unipow}
\title{Ladder of powers transformation on a single vector}
\usage{
eda_unipow(
  x,
  p = c(2, 1, 1/2, 0.33, 0, -0.33, -1/2, -1, -2),
  tukey = TRUE,
  bins = 5,
  cex.main = 1.3,
  col = "#DDDDDD",
  border = "#AAAAAA",
  title = "Re-expressed data via ladder of powers",
  ...
)
}
\arguments{
\item{x}{vector}

\item{p}{vector of powers}

\item{tukey}{if TRUE (default), apply Tukey's power transformation, if FALSE
adopt Box-Cox transformation.}

\item{bins}{number of bins in the histogram}

\item{cex.main}{histogram title size (assigned to each histogram plot)}

\item{col}{histogram fill color}

\item{border}{histogram border color}

\item{title}{Overall plot title (set to NULL for no title)}

\item{...}{other parameters passed to the graphics::hist function.}
}
\description{
\code{eda_unipow} re-expresses a vector on the ladder of powers and plots
 the results using a histogram and a density function. Either the Tukey or Box-Cox
 transformation is used in computing the re-expressed values.
}
\details{
The output is a lattice of descriptive plots showing the transformed data
 across different powers.\cr
}
\examples{
data(mtcars)
eda_unipow(mtcars$mpg, bins=6)
}
\references{
Exploratory Data Analysis, by John Tukey
}
