% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_unipow.R
\name{eda_unipow}
\alias{eda_unipow}
\title{Ladder of powers transformation on a single vector}
\usage{
eda_unipow(
  x,
  p = c(2, 1, 1/2, 0.33, 0, -0.33, -1/2, -1, -2),
  tukey = FALSE,
  bins = 5,
  cex.main = 1.3,
  col = "#DDDDDD",
  border = "#AAAAAA",
  title = "Re-expressed data via ladder of powers",
  ...
)
}
\arguments{
\item{x}{Vector}

\item{p}{Vector of powers}

\item{tukey}{If TRUE, apply Tukey's power transformation. If FALSE
adopt Box-Cox transformation}

\item{bins}{Number of histogram bins}

\item{cex.main}{Histogram title size (assigned to each histogram plot)}

\item{col}{Histogram fill color}

\item{border}{Histogram border color}

\item{title}{Overall plot title (set to NULL for no title)}

\item{...}{Other parameters passed to the graphics::hist function.}
}
\value{
{No return value}
}
\description{
\code{eda_unipow} re-expresses a vector on the ladder of powers
  and plots the results using a histogram and a density function. Either the
  Tukey or Box-Cox transformation is used in computing the re-expressed
  values.
}
\details{
The output is a lattice of descriptive plots showing the transformed data
across different powers.\cr
}
\examples{
data(mtcars)
eda_unipow(mtcars$mpg, bins=6, tukey = TRUE)
}
\references{
\itemize{
    \item Tukey, John W. 1977. Exploratory Data Analysis. Addison-Wesley.}
}
