% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_vd.R
\name{eda_vd}
\alias{eda_vd}
\title{Variability decomposition plot}
\usage{
eda_vd(
  dat,
  y = NULL,
  x = NULL,
  stat = median,
  p = 1L,
  tukey = FALSE,
  base = exp(1),
  ...
)
}
\arguments{
\item{dat}{A model of type \code{eda_pol}, \code{eda_npol}, \code{eda_lm},
or \code{lm}, or a dataframe with a response variable, \code{y}, and a
categorical variable, \code{x}.}

\item{y}{response (continuous) variable if \code{dat} is a dataframe,
\code{NULL} otherwise.}

\item{x}{categorical variable if \code{dat} is a dataframe,
\code{NULL} otherwise.}

\item{stat}{statistical function used to fit \code{y} by \code{x} if
\code{dat} is a dataframe, ignored otherwise.}

\item{p}{Power transformation to apply to univariate data. Ignored if
\code{dat} is not a dataframe.}

\item{tukey}{Boolean determining if a Tukey transformation should be adopted
(TRUE) or if a Box-Cox transformation should be adopted (FALSE). Ignored if
\code{dat} is not a dataframe.}

\item{base}{Base used with the \code{log()} function if \code{p = 0}.}

\item{...}{
  Arguments passed on to \code{\link[=.eda_plot_vardecomp]{.eda_plot_vardecomp}}
  \describe{
    \item{\code{response}}{A character string specifying the name of the response
variable column in the \code{dat} data frame.}
    \item{\code{type}}{A character string specifying the type of plot to generate.
Must be either \code{"boxpnt"} or \code{"box"}.}
    \item{\code{input}}{A character string. \code{"reg"} = bivariate model input.
\code{"nway"} = univariate model or N-way table input.}
    \item{\code{eff}}{A list of effect values. Required when \code{input = "nway"}.}
    \item{\code{rotate}}{Logical. If \code{TRUE}, rotates the plot orientation.}
    \item{\code{padding}}{Numeric. Controls padding for plot limits.}
    \item{\code{show.resp}}{Logical. If \code{TRUE}, includes a boxplot for the response
variable.}
    \item{\code{outliers}}{Logical. If \code{TRUE}, outliers are displayed in boxplots.}
    \item{\code{label}}{Logical. Controls whether labels are displayed.}
    \item{\code{order}}{Logical. Controls ordering (likely of factors or effects).}
    \item{\code{cex.txt}}{Numeric. Controls text size.}
    \item{\code{lim}}{Numeric. Explicit limits for the plot axes.}
    \item{\code{overlap}}{Character. Specifies how to handle overlapping points,
must be one of \code{"stack"}, \code{"overplot"}, or \code{"jitter"}.}
    \item{\code{pch}}{Point symbol type. Only applicable if \code{type = "boxpnt"}.}
    \item{\code{p.col}}{Point border color. Only applicable if \code{type = "boxpnt"}.}
    \item{\code{p.fill}}{Point fill color. Only applicable if \code{type = "boxpnt"}.}
    \item{\code{size}}{Point size. Only applicable if \code{type = "boxpnt"}.}
    \item{\code{alpha}}{Transparency level for points (0 = transparent, 1 = opaque).}
    \item{\code{grey}}{Numeric. Controls grayscale coloring for plot elements
and axes.}
  }}
}
\value{
A plot
}
\description{
Generate plots that visualize the fit/effect and residuals from \code{eda_*} objects.
}
\examples{
# Compare regression model residuals to fit
M0 <- lm(mpg ~ hp + cyl, mtcars)
eda_vd(M0)

# By default, points sharing a identical value are "stacked"
# To jitter:
eda_vd(M0, overlap = "jitter")

# To overplot:
eda_vd(M0, overlap = "overplot")

# To represent fit using a boxplot
eda_vd(M0, type = "box")

# Decompose variability in response variable for a univariate dataset.
# Add labels to each level.
eda_vd(chickwts, weight, feed, label = TRUE)

# Add response variable (bisque colored boxplot)
eda_vd(chickwts, weight, feed, label = TRUE, show.resp = TRUE)

}
