% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eda_lm.R
\name{plot.eda_lm}
\alias{plot.eda_lm}
\title{Residuals plot}
\usage{
\method{plot}{eda_lm}(
  x,
  type = "rd",
  xlab = NULL,
  ylab = NULL,
  grey = 0.7,
  pch = 21,
  equal = TRUE,
  p.col = "grey50",
  alpha = 0.7,
  p.fill = "grey80",
  size = 0.8,
  loess = TRUE,
  loe.col = rgb(0.3, 0.3, 1, 1),
  loess.d = list(family = "symmetric", span = 0.7, degree = 1),
  ...
)
}
\arguments{
\item{x}{Object of class \code{eda_lm}.}

\item{type}{Type of residuals plot. Choice between residuals-fit (\code{rf})
and residuals-dependence (\code{rd}) plots.}

\item{xlab}{X label for output plot.}

\item{ylab}{Y label for output plot.}

\item{grey}{Grey level to apply to plot elements (0 to 1 with 1 = black).}

\item{pch}{Point symbol type.}

\item{equal}{Boolean determining if axes lengths should match (i.e. square
plot).}

\item{p.col}{Color for point symbol.}

\item{alpha}{Point transparency (0 = transparent, 1 = opaque). Only applicable
if \code{rgb()} is not used to define point colors.}

\item{p.fill}{Point fill color passed to \code{bg} (Only used for \code{pch}
ranging from 21-25).}

\item{size}{Point size (0-1).}

\item{loess}{Boolean indicating if a loess should be fitted to the plot.
function. A robust loess is used by default.}

\item{loe.col}{LOESS curve color.}

\item{loess.d}{A list of arguments passed to the \code{loess.smooth}
function. A robust loess is used by default.}

\item{...}{Note used.}
}
\value{
Does not return a value.
}
\description{
Generate residuals vs fitted or residuals vs dependence plot from
an \code{eda_lm} class object
}
\details{
The function generates a scatter plot of residuals vs dependence or
residuals vs fitted values plot from a model of class \code{eda_lm}. A loess
line is fitted to the data. By default, a robust loess is adopted with the
\code{"symmetric"} family.
}
\examples{
M1  <- eda_lm(age_height, Months, Height)
plot(M1) # Residual-dependence plot
plot(M1, type = "rf") # Residual-fit plot
}
