% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eda_npol.R
\name{plot.eda_npol}
\alias{plot.eda_npol}
\title{Plot method for eda_npol objects}
\usage{
\method{plot}{eda_npol}(
  x,
  plot = "effects",
  reg = TRUE,
  margin = "residuals",
  legend = TRUE,
  legend.pos = "bottomright",
  legend.inset = 0.03,
  ...
)
}
\arguments{
\item{x}{An object of class \code{eda_npol}.}

\item{plot}{A character string specifying the type of plot to generate.
\describe{
\item{\code{"effects"} (default)}{Generates a variability decomposition
plot showing residuals and factor effects.}
\item{\code{"diagnostic"}}{Generates a diagnostic plot. Its behavior is
controlled by the \code{margin} argument.}
}}

\item{reg}{Logical. If \code{TRUE}, fits a linear regression line to the
diagnostic plot. Only used when \code{plot = "diagnostic"}. This is disabled
when \code{margin = "all"}. Defaults to \code{FALSE}.}

\item{margin}{Character string. Only used when \code{plot = "diagnostic"}. Specifies which
diagnostic plot to generate. Can be the name of a two-way interaction
(e.g., \code{"Load:Length"}), \code{"residuals"} (the default), or \code{"all"}
to overlay all two-way interaction diagnostics.}

\item{legend}{Logical. If \code{TRUE}, legend is added to plot when \code{margin = "all"}.}

\item{legend.pos}{The position of the legend when \code{margin = "all"}. Can be
\code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"},
\code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"} and \code{"center"}.
Defaults to \code{"topright"}.}

\item{legend.inset}{The amount of inset for the legend from the plot border
when \code{margin = "all"}. Defaults to \code{0.03}.}

\item{...}{Additional arguments passed to internal plotting functions.}
}
\value{
A plot is generated on the active graphics device. The function
itself returns \code{NULL}.
}
\description{
Generates either a variability decomposition plot of factor
effects or a diagnostic plot for an object of class \code{eda_npol}.
}
\details{
This method generates two types of plots for \code{eda_npol} objects:

\strong{1. Variability Decomposition Plot (\code{plot = "effects"})} \cr
Visualizes residuals and the spread of all fitted effects.
Calls \code{\link{.eda_plot_vardecomp}}.

\strong{2. Diagnostic Plot (\code{plot = "diagnostic"})} \cr
When used with an object from \code{eda_npol}, this plot's behavior is
controlled by the \code{margin} argument.
Calls \code{\link{.eda_plot_xy}}.
\itemize{
\item \code{margin = "residuals"}: Plots final residuals against their
n-way comparison values to diagnose higher-order non-additivity.
\item \code{margin = "FactorA:FactorB"}: Plots the fitted two-way
interaction effects against their specific comparison values.
\item \code{margin = "all"}: Overlays the diagnostic plots for all
two-way interactions onto a single graph, with each interaction
represented by a different color and symbol.
}
For a main-effect only model, this generates a single
diagnostic plot of residuals versus a composite comparison value.
}
\examples{

# Main effect median polish (i.e. no interaction)
M1 <- eda_npol(yarn, Cycles, Load, Length, Amplitude)
plot(M1) # Plot effect values and residuals
plot(M1, plot = "diagnostic") # Plot residuals vs comparison value

# Full effect median polish (i.e. include two-way interactions)
M2 <- eda_npol(yarn, Cycles, Load, Length, Amplitude, max_order = 2)
plot(M2, plot = "diagnostic") # Plot residuals vs higher-order CV

# Overlay all two-way interaction diagnostics
plot(M2, plot = "diagnostic", margin = "all")

# Generate the diagnostic plot for a specific two-way interaction
plot(M2, plot = "diagnostic", margin = "Load:Length", reg = TRUE)

# Generate side-by-side diagnostic plots for all two-way interactions
numplots <- length(M2$cv) - 1
nameplots <- names(M2$cv)[-(numplots+1)]
nc <- ceiling(sqrt(numplots))      # number of columns
nr <- ceiling(numplots / nc)       # number of row
OP <- par(mfrow=c(nr,nc))
invisible(sapply(nameplots, \(x) plot(M2, plot="diagnostic", margin = x, reg=TRUE)))
par(OP)

}
\seealso{
\code{\link{eda_npol}}, \code{\link{.eda_plot_vardecomp}}, \code{\link{.eda_plot_xy}}
}
