% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eda_polish.R
\name{plot.eda_polish}
\alias{plot.eda_polish}
\title{Plot method for \code{eda_polish} objects}
\usage{
\method{plot}{eda_polish}(
  x,
  plot = "residuals",
  add.cv = FALSE,
  k = NULL,
  col.quant = FALSE,
  colpal = "RdYlBu",
  colrev = TRUE,
  col.eff = TRUE,
  col.com = TRUE,
  adj.mar = TRUE,
  res.size = 1,
  row.size = 1,
  col.size = 1,
  round = 2,
  res.txt = TRUE,
  label.txt = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{eda_polish}, as produced by \code{eda_pol}.}

\item{plot}{A character string specifying the type of plot to generate. Must
be one of \code{"residuals"}, \code{"cv"}, \code{"diagnostic"}, or
\code{"effects"}.}

\item{add.cv}{Logical. If \code{TRUE}, adds kCV (k-fold cross-validation) values
to the residuals plot. Requires specifying \code{k}.}

\item{k}{Numeric. The value of k for kCV. Required when \code{add.cv ==TRUE}.}

\item{col.quant}{Logical. If \code{TRUE}, adopt a quantile color classification
break.}

\item{colpal}{A character string specifying the color palette to use for
coloring plots. Must be a name listed in \code{hcl.pals()}. Defaults to
\code{"RdYlBu"}.}

\item{colrev}{Logical. If \code{TRUE}, reverses the order of colors in the
palette.}

\item{col.eff}{Logical. Controls coloring for effects plots.}

\item{col.com}{Logical. Controls coloring for common effects.}

\item{adj.mar}{Logical. If \code{FALSE}, adjusts graphical margins using
\code{par(mar = c(1.5, 1.5, 1.5, 1.5))}.}

\item{res.size}{Numeric. Controls the size of residual points or text.}

\item{row.size}{Numeric. Controls row element size.}

\item{col.size}{Numeric. Controls column element size.}

\item{round}{Numeric. Number of digits to round values for display (e.g., in
tables).}

\item{res.txt}{Logical. If \code{TRUE}, displays residual text.}

\item{label.txt}{Logical. If \code{TRUE}, displays labels text.}

\item{...}{
  Arguments passed on to \code{\link[=.eda_plot_vardecomp]{.eda_plot_vardecomp}}
  \describe{
    \item{\code{dat}}{A data frame in long form, containing the data to be plotted.}
    \item{\code{response}}{A character string specifying the name of the response
variable column in the \code{dat} data frame.}
    \item{\code{type}}{A character string specifying the type of plot to generate.
Must be either \code{"boxpnt"} or \code{"box"}.}
    \item{\code{input}}{A character string. \code{"reg"} = bivariate model input.
\code{"nway"} = univariate model or N-way table input.}
    \item{\code{eff}}{A list of effect values. Required when \code{input = "nway"}.}
    \item{\code{rotate}}{Logical. If \code{TRUE}, rotates the plot orientation.}
    \item{\code{padding}}{Numeric. Controls padding for plot limits.}
    \item{\code{show.resp}}{Logical. If \code{TRUE}, includes a boxplot for the response
variable.}
    \item{\code{outliers}}{Logical. If \code{TRUE}, outliers are displayed in boxplots.}
    \item{\code{label}}{Logical. Controls whether labels are displayed.}
    \item{\code{order}}{Logical. Controls ordering (likely of factors or effects).}
    \item{\code{cex.txt}}{Numeric. Controls text size.}
    \item{\code{lim}}{Numeric. Explicit limits for the plot axes.}
    \item{\code{overlap}}{Character. Specifies how to handle overlapping points,
must be one of \code{"stack"}, \code{"overplot"}, or \code{"jitter"}.}
    \item{\code{pch}}{Point symbol type. Only applicable if \code{type = "boxpnt"}.}
    \item{\code{p.col}}{Point border color. Only applicable if \code{type = "boxpnt"}.}
    \item{\code{p.fill}}{Point fill color. Only applicable if \code{type = "boxpnt"}.}
    \item{\code{size}}{Point size. Only applicable if \code{type = "boxpnt"}.}
    \item{\code{alpha}}{Transparency level for points (0 = transparent, 1 = opaque).}
    \item{\code{grey}}{Numeric. Controls grayscale coloring for plot elements
and axes.}
    \item{\code{title}}{Plot title. If title is to be omitted, set to \code{NULL}.}
  }}
}
\value{
Produces a plot. The returned value may vary depending on the
\code{plot} type.

\itemize{
\item \href{../articles/polish.html}{Median polish article}}
}
\description{
This is an S3 method for plotting objects of class \code{eda_polish}. It allows
visualization of different aspects of the analysis, including residuals,
cross-validation values, diagnostic plots, or effects.
}
\examples{
# Generate median polish object
M <- eda_pol(inf_mort, region, edu, perc, plot = FALSE)

# Generate residuals plot
plot(M)

# Generate effects plot
plot(M, plot = "effects", label = TRUE)

# Generate diagnostic plot
plot(M, plot = "diagnostic")

# Add a robust regression line
plot(M, plot = "diagnostic", reg = TRUE, robust = TRUE)

}
