% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eda_polish.R
\name{plot.eda_polish}
\alias{plot.eda_polish}
\title{Plot eda_polish tables or diagnostic plots}
\usage{
\method{plot}{eda_polish}(
  x,
  type = "residuals",
  k = 1,
  col.quant = FALSE,
  colpal = "RdYlBu",
  adj.mar = FALSE,
  res.size = 1,
  row.size = 1,
  col.size = 1,
  res.txt = TRUE,
  label.txt = TRUE,
  ...
)
}
\arguments{
\item{x}{A list of class eda_polish}

\item{type}{Plot type. One of three: "residuals", "cv" or "diagnostic".}

\item{k}{If cv values are to be plotted, define the k parameter}

\item{col.quant}{Boolean determining if a quantile classification scheme should be used}

\item{colpal}{Color palette to adopt}

\item{adj.mar}{Boolean determining if margin width needs to accomodate labels}

\item{res.size}{Size of residual values in plot [0-1]}

\item{row.size}{Size of row effect values in plot [0-1]}

\item{col.size}{Size of column effect values in plot [0-1]}

\item{res.txt}{Boolean determining if values should be added to plot}

\item{label.txt}{Boolean determining if margin and column labels should be plotted}

\item{...}{Arguments to be passed to subsequent methods}
}
\description{
\code{eda_pol} A plot method for lists of eda_polish class.
}
\details{
The function plots a polish table of residuals or CV values. I will also generate
 a diagnostic plot if \code{type} is set to \code{diagnostic}
}
