% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eda_polish.R
\name{plot.eda_polish}
\alias{plot.eda_polish}
\title{Plot eda_polish tables or diagnostic plots}
\usage{
\method{plot}{eda_polish}(
  x,
  type = "residuals",
  k = 1,
  col.quant = FALSE,
  colpal = "RdYlBu",
  col.eff = TRUE,
  adj.mar = FALSE,
  res.size = 1,
  row.size = 1,
  col.size = 1,
  res.txt = TRUE,
  label.txt = TRUE,
  ...
)
}
\arguments{
\item{x}{A list of class eda_polish}

\item{type}{Plot type. One of three: "residuals", "cv" or "diagnostic".}

\item{k}{If cv values are to be plotted, define the k parameter}

\item{col.quant}{Boolean indicating if a quantile classification scheme should be used}

\item{colpal}{Color palette to adopt}

\item{col.eff}{Boolean indicating if effects should contribute to color gradient}

\item{adj.mar}{Boolean indicating if margin width needs to accommodate labels}

\item{res.size}{Size of residual values in plot [0-1]}

\item{row.size}{Size of row effect values in plot [0-1]}

\item{col.size}{Size of column effect values in plot [0-1]}

\item{res.txt}{Boolean indicating if values should be added to plot}

\item{label.txt}{Boolean indicating if margin and column labels should be plotted}

\item{...}{Arguments to be passed to subsequent methods}
}
\description{
\code{eda_pol} A plot method for lists of eda_polish class.
}
\details{
The function plots a polish table of residuals or CV values. I will also generate
 a diagnostic plot if \code{type} is set to \code{diagnostic}
}
\examples{
# Create dataset
df <- data.frame(region =  rep( c("NE", "NC", "S", "W"), each = 5),
edu = rep( c("ed8", "ed9to11", "ed12", "ed13to15", "ed16"), 4),
perc = c(25.3, 25.3, 18.2, 18.3, 16.3, 32.1, 29, 18.8,
        24.3, 19, 38.8, 31, 19.3, 15.7, 16.8, 25.4, 21.1, 20.3, 24, 17.5))

# Generate median polish output
out <- eda_pol(df, row = "region", col = "edu", val = "perc", plot = FALSE)

# Plot table
plot(out, type = "residuals")


# Plot table using CV values
plot(out, type = "cv")

# Generate diagnostic plot
plot(out, type = "diagnostic")
}
