% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyedar.R
\name{tukeyedar}
\alias{tukeyedar}
\title{Tukey inspired exploratory data analysis functions}
\description{
This packages hosts a small set of Tukey inspired functions for use
in exploring datasets in a robust manner.
}
