% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyedar.R
\docType{package}
\name{tukeyedar}
\alias{tukeyedar-package}
\alias{tukeyedar}
\title{Tukey inspired exploratory data analysis functions}
\description{
This packages hosts a small set of Tukey inspired functions for use
in exploring datasets in a robust manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mgimond.github.io/tukeyedar/}
  \item Report bugs at \url{https://github.com/mgimond/tukeyedar/issues/}
}

}
\author{
\strong{Maintainer}: Manuel Gimond \email{mgimond@gmail.com} (\href{https://orcid.org/0000-0002-3687-2316}{ORCID})

}
